% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADAE}
\alias{ADAE}
\title{Example ADAE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 55 columns.
}
\usage{
ADAE
}
\description{
Data frame imported from the \href{https://github.com/cdisc-org/sdtm-adam-pilot-project}{CDISC SDTM/ADaM Pilot Project}
}
\keyword{datasets}
