% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adam}
\alias{adam}
\alias{ADSL}
\alias{ADAE}
\alias{ADTTE}
\alias{ADLB}
\title{Example ADaM Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 26 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5784 rows and 46 columns.
}
\usage{
ADSL

ADAE

ADTTE

ADLB
}
\description{
Data frame imported from the \href{https://github.com/cdisc-org/sdtm-adam-pilot-project}{CDISC SDTM/ADaM Pilot Project}
}
\keyword{datasets}
