% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_box_card}
\alias{add_box_card}
\title{Add box card}
\usage{
add_box_card(
  data = NULL,
  name = "card_name",
  title = "Card Title",
  desc = "Card Description",
  value = "Card Value",
  succ_icon = "black-tie",
  icon_class = "text-info",
  is_perc = FALSE,
  is_url = FALSE
)
}
\arguments{
\item{data}{a data.frame}
}
\description{
Pretty simple function that makes new card data creation really simple, and
adds small cheat to create the table if it doesn't already exist
}
