% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADTTE}
\alias{ADTTE}
\title{Example ADTTE}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 26 columns.
}
\usage{
ADTTE
}
\description{
Data frame imported from the \href{https://github.com/cdisc-org/sdtm-adam-pilot-project}{CDISC SDTM/ADaM Pilot Project}
}
\keyword{datasets}
