% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fmt_fn.R
\name{apply_fmt_fn}
\alias{apply_fmt_fn}
\title{Apply Formatting Functions}
\usage{
apply_fmt_fn(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}
}
\value{
an ARD data frame of class 'card'
}
\description{
Apply the formatting functions to each of the raw statistics.
Function aliases are converted to functions using \code{\link[=alias_as_fmt_fn]{alias_as_fmt_fn()}}.
}
\examples{
ard_continuous(ADSL, variables = "AGE") |>
  apply_fmt_fn()
}
