% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fmt_fun.R
\name{apply_fmt_fun}
\alias{apply_fmt_fun}
\title{Apply Formatting Functions}
\usage{
apply_fmt_fun(x, replace = FALSE)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}

\item{replace}{(scalar \code{logical})\cr
logical indicating whether to replace values in the \code{'stat_fmt'} column (if present).
Default is \code{FALSE}.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Apply the formatting functions to each of the raw statistics.
Function aliases are converted to functions using \code{\link[=alias_as_fmt_fun]{alias_as_fmt_fun()}}.
}
\examples{
ard_summary(ADSL, variables = "AGE") |>
  apply_fmt_fun()
}
