% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical.R
\name{ard_categorical}
\alias{ard_categorical}
\title{Categorical ARD Statistics}
\usage{
ard_categorical(data, variables, by = NULL, strata = NULL, denominator = NULL)
}
\arguments{
\item{data}{a data frame}

\item{variables}{columns to include in summaries. Default is \code{everything()}.}

\item{by, strata}{columns to by/stratified by for tabulation.
Arguments are similar, but with an important distinction:

\code{by}: results are tabulated by \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.

\code{strata}: results are tabulated by \strong{all \emph{observed} combinations} of the
columns specified.

Arguments may be used in conjunction with one another.}

\item{denominator}{Specify this \emph{optional} argument to change the denominator,
e.g. the \code{"N"} statistic. Default is \code{NULL}. See below for details.}
}
\value{
a data frame
}
\description{
Compute Analysis Results Data (ARD) for categorical summary statistics.
}
\section{Denominators}{

By default, the \code{ard_categorical()} function returns the statistics \code{"n"} and \code{"N"},
where little \code{"n"} are the counts for the variable levels, and \code{"N"} is
the number of non-missing observations. The default calculation for the
percentage is merely \code{p = n/N}.

However, it is sometimes necessary to provide a different \code{"N"} to use
as the denominator in this calculation. For example, in a calculation
of the rates of various observed adverse events, you may need to update the
denominator to the number of enrolled subjects.

In such cases, use the \code{denominator} argument to specify a new definition
of \code{"N"}, and subsequently \code{"p"}.
The argument expects a data frame, and the data frame must include the columns
specified in \code{ard_categorical(by=)} (strata columns are not considered).
The updated \code{N} and \code{length} elements will be updated to be calculated as
the number of rows in each combination of the \code{by} variables.

Take an example where we need to update the denominator to be subjects enrolled
in a trial, e.g. tabulating the number of AEs that occurred within an SOC
where some subjects may not have experienced an AE and would not be represented
in the ADAE data set. All patients appear in ADSL, however.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ard_categorical(
  data =
    ADAE |>
      dplyr::left_join(ADSL[c("USUBJID", "ARM")], by = "USUBJID") |>
      dplyr::filter(AOCCSFL \%in\% "Y"),
  by = "ARM",
  variables = "AESOC",
  denominator = ADSL
) |>
  flatten_ard()
#> # A tibble: 141 x 9
#>    group1 group1_level variable variable_level    stat_name stat_label statistic
#>    <chr>  <chr>        <chr>    <chr>             <chr>     <chr>      <chr>    
#>  1 ARM    Placebo      AESOC    CARDIAC DISORDERS n         n          12       
#>  2 ARM    Placebo      AESOC    CARDIAC DISORDERS p         \%          0.139534~
#>  3 ARM    Placebo      AESOC    CONGENITAL, FAMI~ n         n          0        
#>  4 ARM    Placebo      AESOC    CONGENITAL, FAMI~ p         \%          0        
#>  5 ARM    Placebo      AESOC    EAR AND LABYRINT~ n         n          1        
#>  6 ARM    Placebo      AESOC    EAR AND LABYRINT~ p         \%          0.011627~
#>  7 ARM    Placebo      AESOC    EYE DISORDERS     n         n          2        
#>  8 ARM    Placebo      AESOC    EYE DISORDERS     p         \%          0.023255~
#>  9 ARM    Placebo      AESOC    GASTROINTESTINAL~ n         n          17       
#> 10 ARM    Placebo      AESOC    GASTROINTESTINAL~ p         \%          0.197674~
#> # i 131 more rows
#> # i 2 more variables: warning <chr>, error <chr>
}\if{html}{\out{</div>}}
}

\examples{
ard_categorical(ADSL, by = "ARM", variables = "AGEGR1") |>
  flatten_ard()
}
