% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_compare.R
\name{ard_compare}
\alias{ard_compare}
\title{Compare ARDs}
\usage{
ard_compare(x, y, key_columns = NULL)
}
\arguments{
\item{x}{(\code{card})\cr
first ARD to compare.}

\item{y}{(\code{card})\cr
second ARD to compare.}

\item{key_columns}{(\code{character} or \code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
optional specification of column names identifying unique records. Supply a
character vector or tidyselect expression. When \code{NULL}, grouping columns
along with \code{"variable"}, \code{"variable_level"}, and \code{"stat_name"} are used.
If the ARDs do not have identical primary key sets, the comparison falls
back to the shared primary key columns provided they uniquely identify rows
in both inputs.}
}
\value{
a named list of data frames containing key columns with the
corresponding values from both ARDs for rows where the column values do not
match. The list contains entries for the \code{stat}, \code{fmt_fun}, \code{warning}, and
\code{error} columns (the \code{fmt_fun} component is labeled \code{fmt_fn} for
compatibility with previous naming conventions).
}
\description{
\code{ard_compare()} compares the \code{stat}, \code{fmt_fun}, \code{warning}, and \code{error}
columns of two ARDs row-by-row using a shared set of key columns. Rows where
the column values differ are returned.
}
\details{
When both ARDs retain metadata about the environment that created them (for
example via an attribute storing an environment object), the function ensures
that the environments match before comparing values.
}
\examples{
ard_base <- ard_summary(ADSL, variables = AGE)
ard_modified <- ard_summary(dplyr::mutate(ADSL, AGE = AGE + 1), variables = AGE)

ard_compare(ard_base, ard_modified)$stat

}
