% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{ard_continuous}
\alias{ard_continuous}
\title{Continuous ARD Statistics}
\usage{
ard_continuous(
  data,
  variables,
  by = NULL,
  statistics = everything() ~ continuous_variable_summary_fns()
)
}
\arguments{
\item{data}{a data frame}

\item{variables}{columns to include in summaries.}

\item{by}{columns to compute statistics by. Default are the columns
returned by \code{dplyr::group_vars(data)}.}

\item{statistics}{a named list of functions, a list of formulas,
or a single formula where the list element is a function (or the RHS of
a formula),
e.g. \verb{list(mpg = list(mean = \\(x) mean(x, na.rm = TRUE)))}}
}
\value{
a data frame
}
\description{
Compute Analysis Results Data (ARD) for simple continuous summary statistics.
}
\examples{
ard_continuous(ADSL, by = "ARM", variables = "AGE") |>
  flatten_ard()
}
