% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_filter.R
\name{ard_filter}
\alias{ard_filter}
\title{Filter Stacked Hierarchical ARDs}
\usage{
ard_filter(x, filter)
}
\arguments{
\item{x}{(\code{card})\cr
A stacked hierarchical ARD of class \code{'card'} created using \code{\link[=ard_stack_hierarchical]{ard_stack_hierarchical()}}.}

\item{filter}{(\code{expression})\cr an expression that is used to filter rows of the hierarchical ARD. See details in the
Filtering section below.}
}
\value{
an ARD data frame of class 'card'
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr

This function is used to filter stacked hierarchical ARDs.
}
\details{
The \code{filter} argument can be used to filter out rows of a hierarchical ARD which do not meet the requirements
provided as an expression. Rows can be filtered on the values of any of the possible statistics (\code{n}, \code{p}, and \code{N})
provided they are included at least once in the ARD. For each entry that does not meet the filtering requirement,
all statistics corresponding to that entry will be removed from the ARD. In addition to filtering on individual
statistic values, filters can be applied across the hierarchical row (i.e. across all \code{by} variable values) by using
aggregate functions such as \code{sum()} and \code{mean()}.

Some examples of possible filters:
\itemize{
\item \code{filter = n > 5}
\item \code{filter = n == 2 & p < 0.05}
\item \code{filter = sum(n) > 4}
\item \code{filter = mean(n) > 4 | n > 3}
}
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ard_stack_hierarchical(
  ADAE,
  variables = c(AESOC, AEDECOD),
  by = TRTA,
  denominator = ADSL |> dplyr::rename(TRTA = ARM),
  id = USUBJID
) |>
  ard_filter(sum(n) > 3)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=ard_stack_hierarchical]{ard_stack_hierarchical()}}
}
