% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_label.R
\name{ard_label}
\alias{ard_label}
\title{ARD Labels}
\usage{
ard_label(data, label = NULL, include = everything())
}
\arguments{
\item{data}{a data frame}

\item{label}{named list of variable labels, e.g. \code{list(mpg = "MPG")}.
Default is \code{NULL}}

\item{include}{variable to include}
}
\value{
a data frame
}
\description{
Add variable labels to an ARD data frame.
When no label specified and no label has been set for a column,
the column name will be placed in the label statistic.
}
\examples{
df <- dplyr::tibble(var1 = letters, var2 = LETTERS)
attr(df$var1, 'label') <- "Lowercase Letters"

ard_label(df)
}
