% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_regression.R
\name{ard_regression}
\alias{ard_regression}
\title{Regression ARD}
\usage{
ard_regression(model, tidy_fun = NULL, ...)
}
\arguments{
\item{model}{regression model object}

\item{tidy_fun}{a tidier. Default is broom.helpers::tidy_with_broom_or_parameters}

\item{...}{Arguments passed to \code{broom.helpers::tidy_plus_plus()}}
}
\value{
data frame
}
\description{
Function takes a regression model object and converts it to a ARD
structure using the \code{broom.helpers} package.
}
\examples{
lm(hp ~ mpg + factor(cyl), data = mtcars) |>
  ard_regression(add_estimate_to_reference_rows = TRUE)
}
