% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_stratified_categorical.R
\name{ard_stratified_categorical}
\alias{ard_stratified_categorical}
\title{Stratified Categorical ARD Statistics}
\usage{
ard_stratified_categorical(
  data,
  strata,
  variables,
  by = NULL,
  denominator = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{strata}{a single column from \code{data}.}

\item{variables}{columns to include in summaries. Default is \code{everything()}.}

\item{by}{columns to compute statistics by. Default are the columns
returned by \code{dplyr::group_vars(data)}.}

\item{denominator}{Specify this \emph{optional} argument to change the denominator,
e.g. the \code{"N"} statistic. Default is \code{NULL}. See below for details.}
}
\value{
a data frame
}
\description{
Similar to \code{ard_categorical()} except each call to \code{ard_categorical()}
is called within each stratum level.
The stratum variable and levels are returned in the ARD in columns
\code{"group##"} and \code{"group##_level"}
}
\examples{
ard_stratified_categorical(
  data = ADAE |> dplyr::filter(AOCCPFL \%in\% "Y"),
  strata = "AESOC",
  variables = "AEDECOD",
  denominator = ADSL
)
}
