% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_tabulate_rows.R
\name{ard_tabulate_rows}
\alias{ard_tabulate_rows}
\title{Row Tabulate ARD}
\usage{
ard_tabulate_rows(
  data,
  colname = "..row_count..",
  by = dplyr::group_vars(data),
  strata = NULL,
  fmt_fun = NULL
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{colname}{(\code{string})\cr
name of the column that will be returned along with the row tabulation.}

\item{by, strata}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to use for grouping or stratifying the table output.
Arguments are similar, but with an important distinction:

\code{by}: results are tabulated by \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.

\code{strata}: results are tabulated by \strong{all \emph{observed} combinations} of the
columns specified.

Arguments may be used in conjunction with one another.}

\item{fmt_fun}{(\code{\link[=syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character()))}.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Tabulate the number of rows in a data frame.
}
\examples{
ard_tabulate_rows(ADSL, by = TRTA)
}
