% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_proportion_ci.R
\name{calc_prop_ci}
\alias{calc_prop_ci}
\alias{.calc_prop_ci_wald}
\title{Helper Functions for Calculating Proportion Confidence Intervals}
\usage{
.calc_prop_ci_wald(x, conf.level, correct = FALSE)
}
\arguments{
\item{correct}{(\code{logical})\cr apply continuity correction.}
}
\value{
Confidence interval of a proportion.
}
\description{
Functions to calculate different proportion confidence intervals for use in \code{ard_proportion()}.
}
\section{Functions}{
\itemize{
\item \code{.calc_prop_ci_wald()}: Calculates the Wald interval by following the usual textbook definition
for a single proportion confidence interval using the normal approximation.

}}
\keyword{internal}
