% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CardClass.R
\docType{class}
\name{Card-class}
\alias{Card-class}
\title{Card}
\description{
Regular Playing Cards
}
\section{Slots}{

\describe{
\item{\code{suit}}{Determines the suit of the card: Spade, Club, Heart, or Diamond}

\item{\code{value}}{Determines the card's value 2-14: 11 = Jack, 12 = Queen, 13 = King, 14 = Ace}
}}

