% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardConstructor.R
\name{card}
\alias{card}
\title{Card Constructor Function}
\usage{
card(suit, value)
}
\arguments{
\item{suit}{Determines the suit of the card: Spade, Club, Heart, or Diamond}

\item{value}{Determines the card's value 2-14: 11 = Jack, 12 = Queen, 13 = King, 14 = Ace}
}
\description{
This function constructs a card based off of the user's inputs.
}
