% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal-package.R
\docType{package}
\name{cardinal-package}
\alias{cardinal}
\alias{cardinal-package}
\title{cardinal: CDISC Aanalys Results Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
