% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cardplot}
\alias{cardplot}
\alias{cardplot,ballgown-method}
\title{make an awesome plot of a card!}
\usage{
plot(x, y, ...)

\S4method{plot}{Card}(x)
}
\arguments{
\item{x}{a Card object}
}
\description{
make an awesome plot of a card!
}

