% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{check_binary}
\alias{check_binary}
\title{Check Binary}
\usage{
check_binary(x, arg_name = caller_arg(x), call = parent.frame())
}
\arguments{
\item{x}{a vector}

\item{call}{call environment}
}
\value{
invisible
}
\description{
Checks if a column in a data frame is binary,
that is, if the column is class \verb{<logical>} or
\verb{<numeric/integer>} and coded as \code{c(0, 1)}
}
\keyword{internal}
