% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{check_length}
\alias{check_length}
\title{Check Length}
\usage{
check_length(
  x,
  arg_name = caller_arg(x),
  length = 1L,
  msg = "The {.arg {arg_name}} argument must be length {.val {length}}.",
  call = parent.frame()
)
}
\arguments{
\item{x}{object to check}

\item{length}{integer specifying the required length}

\item{msg}{string passed to \code{cli::cli_abort(message=)}}
}
\description{
Check Length
}
\keyword{internal}
