% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{check_range}
\alias{check_range}
\title{Check Range}
\usage{
check_range(
  x,
  range,
  include_bounds = c(FALSE, FALSE),
  arg_name = caller_arg(x),
  msg = paste("The {.arg {arg_name}} argument must be in the interval",
    "{.code {ifelse(include_bounds[1], '[', '(')}{range[1]}, {range[2]}{ifelse(include_bounds[2], ']', ')')}}."),
  call = parent.frame()
)
}
\arguments{
\item{x}{numeric scalar to check}

\item{range}{numeric vector of length two}

\item{include_bounds}{logical of length two indicating whether to allow
the lower and upper bounds}

\item{msg}{string passed to \code{cli::cli_abort(message=)}}
}
\description{
Check Range
}
\keyword{internal}
