% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cardinal.R
\name{construct_cardinal}
\alias{construct_cardinal}
\title{Construct Cardinal Table}
\usage{
construct_cardinal(table_plan, header_plan, hide = c("variable", "header_row"))
}
\arguments{
\item{table_plan}{a table plan data frame}

\item{header_plan}{a header plan named list}

\item{hide}{character vector of columns to hide from output}
}
\value{
a cardinal table
}
\description{
Construct a table of class 'cardinal' from the table plan, header plan, and
other table compnents
}
\examples{
# Construct the ARD
table_ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, include = mpg),
    ard_categorical(mtcars, by = cyl, include = am),
    ard_categorical(mtcars,  include = cyl)
  )

# convert ARD to a cardinal table
table <-
  construct_cardinal(
    table_plan =
      dplyr::bind_rows(
        table_ard |> dplyr::filter(variable \%in\% "mpg") |>  table_plan_simple_continuous(),
        table_ard |> dplyr::filter(variable \%in\% "am") |> table_plan_simple_categorical()
      ),
    header_plan =
      table_ard |>
      dplyr::filter(variable \%in\% "cyl") |>
      header_plan_simple(header = "**{strata} Cylinders  \nN = {n} ({p}\%)**") |>
      utils::modifyList(val = list(label = "**Characteristic**"))
  )
}
