% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cardinal.R
\name{construct_cards}
\alias{construct_cards}
\title{Construct cards Table}
\usage{
construct_cards(table_plan, header_plan, hide = c("variable", "header_row"))
}
\arguments{
\item{table_plan}{a table plan data frame}

\item{header_plan}{a header plan named list}

\item{hide}{character vector of columns to hide from output}
}
\value{
a cards table
}
\description{
Construct a table of class 'cards' from the table plan, header plan, and
other table components
}
\examples{
# Construct the ARD
table_ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, variables = "mpg"),
    ard_categorical(mtcars, by = cyl, variables = "am"),
    ard_categorical(mtcars,  variables = "cyl")
  )

# convert ARD to a cards table
table <-
  construct_cards(
    table_plan =
      dplyr::bind_rows(
        table_ard |> dplyr::filter(variable \%in\% "mpg") |>  table_plan_simple_continuous(),
        table_ard |> dplyr::filter(variable \%in\% "am") |> table_plan_simple_categorical()
      ),
    header_plan =
      table_ard |>
      dplyr::filter(variable \%in\% "cyl") |>
      header_plan_simple(header = "**{group} Cylinders  \nN = {n} ({p}\%)**") |>
      utils::modifyList(val = list(label = "**Characteristic**"))
  )
}
