% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_variable_summary_fns.R
\name{continuous_variable_summary_fns}
\alias{continuous_variable_summary_fns}
\title{List of Continuous Summary Functions}
\usage{
continuous_variable_summary_fns(summaries = NULL)
}
\arguments{
\item{summaries}{a character vector of functions to include in output. Default
is \code{NULL}, which selects all summary functions. Select one or more from
'N', 'mean', 'sd', 'median', 'p25', 'p75', 'min', 'max'.}
}
\value{
named list of summary functions
}
\description{
Returns a list of univariate summary functions.
Some functions include slight modifications to their base equivalents. For
example, the \code{min()} and \code{max()} functions return \code{NA} instead of \code{Inf}
when an empty vector is passed.
}
\examples{
continuous_variable_summary_fns(c("N", "median"))
}
