% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cardinal.R
\name{convert_cardinal}
\alias{convert_cardinal}
\title{Convert Table to Print Engine Class}
\usage{
convert_cardinal(x, engine = "gt")
}
\arguments{
\item{x}{a cardinal table}

\item{engine}{a string indicating the print engine. Default is \code{"gtß"}}
}
\value{
a table object
}
\description{
Convert a 'cardinal' table to any of the print engine classes.
Goal is to support gt, rtables, flextable, huxtable, kableExtra,
kable, and formatted tibbles
}
\examples{
# Construct the ARD
table_ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, include = mpg),
    ard_categorical(mtcars, by = cyl, include = am),
    # TODO: The ARD creation code can by simplified after the categorical
    #       ARD accepts no-by variable specifications
    mtcars |>
      dplyr::mutate(..one.. = 1L) |>
      ard_categorical(by = ..one..,  include = cyl) |>
      dplyr::select(-starts_with("strata"))
  )

# convert ARD to a cardinal table
construct_cardinal(
  plan_table =
    dplyr::bind_rows(
      table_ard |> dplyr::filter(variable \%in\% "mpg") |>  plan_table_simple_continuous(),
      table_ard |> dplyr::filter(variable \%in\% "am") |> plan_table_simple_categorical()
    ),
  plan_header =
    table_ard |>
    dplyr::filter(variable \%in\% "cyl") |>
    plan_header_simple(header = "**{strata} Cylinders  \nN = {n}  ({p}\%)**") |>
    modifyList(val = list(label = gt::md("**Characteristic**")))
) |>
  convert_cardinal()
}
