% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cards.R
\name{convert_cards}
\alias{convert_cards}
\title{Convert Table to Print Engine Class}
\usage{
convert_cards(x, engine = "gt")
}
\arguments{
\item{x}{a cards table}

\item{engine}{a string indicating the print engine. Default is \code{"gtß"}}
}
\value{
a table object
}
\description{
Convert a 'cards' table to any of the print engine classes.
Goal is to support gt, rtables, flextable, huxtable, kableExtra,
kable, and formatted tibbles
}
\examples{
# Construct the ARD
table_ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, variables = "mpg"),
    ard_categorical(mtcars, by = cyl, variables = "am"),
    ard_categorical(mtcars,  variables = "cyl")
  )

# convert ARD to a cards table
construct_cards(
  table_plan =
    dplyr::bind_rows(
      table_ard |> dplyr::filter(variable \%in\% "mpg") |>  table_plan_simple_continuous(),
      table_ard |> dplyr::filter(variable \%in\% "am") |> table_plan_simple_categorical()
    ),
  header_plan =
    table_ard |>
    dplyr::filter(variable \%in\% "cyl") |>
    header_plan_simple(header = "**{group} Cylinders  \nN = {n}  ({p}\%)**") |>
    modifyList(val = list(label = gt::md("**Characteristic**")))
) |>
  convert_cards()
}
\keyword{internal}
