% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck-df.R
\name{deal_hand_df}
\alias{deal_hand_df}
\title{Deal a hand of cards using \code{data.frame}}
\usage{
deal_hand_df(d)
}
\arguments{
\item{d}{a deck created by \code{\link[=new_deck_df]{new_deck_df()}}.}
}
\value{
a \code{data.frame} containing 5 random cards.
}
\description{
Return a five card hand from a deck of cards, as a \code{data.frame} with \code{rank} and \code{suit}.
}
\details{
Uses \code{\link[=sample]{sample()}} to select 5 random cards.
}
\examples{
deck <- new_deck_df()
deal_hand_df(deck)
}
