% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_stat_args.R
\name{default_stat_labels}
\alias{default_stat_labels}
\alias{default_fmt_fns}
\title{Defaults for statistical arguments}
\usage{
default_stat_labels()

default_fmt_fns()
}
\value{
named list of values
}
\description{
-\code{default_stat_labels()} returns a named list of statistics labels
\itemize{
\item \code{default_fmt_fns()} returns a named list of formatting functions to be applied in lieu
of the global default, which is to round to one decimal place.
}
}
\examples{
# stat labels
default_stat_labels()

# formatting functions
default_fmt_fns()

}
