% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_stat_labels.R
\name{default_stat_labels}
\alias{default_stat_labels}
\title{Defaults for Statistical Arguments}
\usage{
default_stat_labels()
}
\value{
named list
}
\description{
Returns a named list of statistics labels
}
\examples{
# stat labels
default_stat_labels()
}
