% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical.R
\name{.ard_categorical_recalc_N}
\alias{.ard_categorical_recalc_N}
\title{Recalculate N and p}
\usage{
.ard_categorical_recalc_N(df_result, denominator, by, variables)
}
\arguments{
\item{df_result}{an ARD data frame}

\item{denominator}{a data frame}

\item{by}{character vector of by variables passed in \code{ard_categorical(by)}}

\item{variables}{vector string of the variables passed in \code{ard_categorical(variables)}}
}
\value{
an ARD data frame
}
\description{
If a user passes \code{ard_categorical(denominator)} the N and p need to be adjusted
to match the data frame in \code{denominator}.
}
\examples{
ard_categorical(
  mtcars,
  variables = cyl
) |>
  cards:::.ard_categorical_recalc_N(
    denominator = rep(list(mtcars), 10) |> dplyr::bind_rows(),
    by = NULL,
    variables = "cyl"
  ) |>
  flatten_ard()
}
\keyword{internal}
