% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_dichotomous.R
\name{.check_dichotomous_values}
\alias{.check_dichotomous_values}
\title{Perform Value Checks}
\usage{
.check_dichotomous_values(data, values, call = caller_env())
}
\arguments{
\item{data}{(\code{data.frame})
a data frame}

\item{values}{(named \code{list})
a named list}

\item{call}{(\code{environment})
an environment. Default is \code{caller_env()}}
}
\description{
Check the validity of the values passed in \code{ard_dichotomous(values)}.
}
\keyword{internal}
