% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ard_conditions.R
\name{.cli_condition_messaging}
\alias{.cli_condition_messaging}
\title{Print Condition Messages Saved in an ARD}
\usage{
.cli_condition_messaging(x, msg_type, condition_type)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}

\item{msg_type}{(\code{string})\cr
message type. Options are \code{"warning"} and \code{"error"}.}
}
\value{
returns invisible if check is successful, throws warning/error messages if not.
}
\description{
Print Condition Messages Saved in an ARD
}
\examples{
ard <- ard_summary(
  ADSL,
  by = ARM,
  variables = AGE
)

cards:::.cli_condition_messaging(ard, msg_type = "error")
}
\keyword{internal}
