% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.default_fmt_fn}
\alias{.default_fmt_fn}
\title{Add Default Formatting Functions}
\usage{
.default_fmt_fn(x)
}
\arguments{
\item{data}{frame with cards structure}
}
\value{
data frame
}
\description{
Add Default Formatting Functions
}
\keyword{internal}
