% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_summary.R
\name{.default_fmt_fun}
\alias{.default_fmt_fun}
\title{Add Default Formatting Functions}
\usage{
.default_fmt_fun(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}
}
\value{
a data frame
}
\description{
Add Default Formatting Functions
}
\examples{
ard <- ard_tabulate(ADSL, by = "ARM", variables = "AGEGR1") |>
  dplyr::mutate(fmt_fun = NA)

cards:::.default_fmt_fun(ard)
}
\keyword{internal}
