% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.default_statistic_formatters}
\alias{.default_statistic_formatters}
\title{Default formatting functions}
\usage{
.default_statistic_formatters()
}
\value{
named list of functions
}
\description{
Global default is to round the statistic to one decimal place for
all other rounding, the default function must be listed below
}
\examples{
cards:::.default_statistic_formatters()
}
\keyword{internal}
