% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ard_stat_string.R
\name{.extract_stat_names}
\alias{.extract_stat_names}
\title{Extract Statistic Names from Pattern}
\usage{
.extract_stat_names(pattern)
}
\arguments{
\item{pattern}{(\code{string})\cr
a glue pattern, e.g. \code{"{n} ({p}\%)"}}
}
\value{
a character vector of the unique names inside curly brackets
}
\description{
Extract Statistic Names from Pattern
}
\examples{
cards:::.extract_stat_names("{n} / {N} ({p}\%)")
}
\keyword{internal}
