% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_proportion_ci.R
\name{.is_binary}
\alias{.is_binary}
\title{Is Binary?}
\usage{
.is_binary(data, variable, call = caller_env())
}
\arguments{
\item{data}{a data frame}

\item{variable}{\verb{<string>}\cr a column name from \code{data}}

\item{call}{call environment}
}
\value{
invisible
}
\description{
Checks if a column in a data frame is binary,
that is, if the column is class \verb{<logical>} or
numeric and coded as \code{c(0, 1)}
}
\keyword{internal}
