% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.is_list_column_of_scalars}
\alias{.is_list_column_of_scalars}
\title{List Column as a Vector Predicate}
\usage{
.is_list_column_of_scalars(x)
}
\arguments{
\item{x}{(\code{any})\cr
column to check}
}
\value{
a logical
}
\description{
A predicate function to check whether a column is a list and can be
represented as a vector.
}
\examples{
cards:::.is_list_column_of_scalars(as.list(1:5))
}
\keyword{internal}
