% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nested_list.R
\name{.one_row_ard_to_nested_list}
\alias{.one_row_ard_to_nested_list}
\title{Convert One Row to Nested List}
\usage{
.one_row_ard_to_nested_list(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card' with one row}
}
\value{
an expression that represents an element of a nested list
}
\description{
Convert One Row to Nested List
}
\examples{
ard_summary(mtcars, variables = mpg) |>
  dplyr::filter(dplyr::row_number() \%in\% 1L) |>
  apply_fmt_fun() |>
  cards:::.one_row_ard_to_nested_list()
}
\keyword{internal}
