% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.process_stat_labels}
\alias{.process_stat_labels}
\title{Process Statistic Labels}
\usage{
.process_stat_labels(stat_labels)
}
\arguments{
\item{stat_labels}{named list}
}
\value{
named list
}
\description{
Process Statistic Labels
}
\examples{
list(AGE = list(c("N", "n") ~ "{n} / {N}")) |>
  cards:::.process_stat_labels()
}
\keyword{internal}
