% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.rename_ard_groups}
\alias{.rename_ard_groups}
\title{Rename ARD Group Variables}
\usage{
.rename_ard_groups(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame}
}
\value{
data frame
}
\description{
This function combines each pair of \code{group} and \code{group_level} columns into a
single column. The \code{group_level} column is renamed according to the value of
the \code{group} column.
}
\examples{
data <- data.frame(group1 = "A", group1_level = "B", group2 = "C", group2_level = "D")

cards:::.rename_ard_groups(data)
}
\keyword{internal}
