% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_hierarchical.R
\name{.rename_last_group_as_variable}
\alias{.rename_last_group_as_variable}
\title{Rename Last Group to Variable}
\usage{
.rename_last_group_as_variable(df_result, by, variables)
}
\arguments{
\item{df_result}{(\code{data.frame})\cr
an ARD data frame of class 'card'}
}
\value{
an ARD data frame of class 'card'
}
\description{
In the \verb{ard_hierarchical*()} functions, the last grouping variable is
renamed to \code{variable} and \code{variable_level} before being returned.
}
\examples{
data <- data.frame(x = 1, y = 2, group1 = 3, group2 = 4)

cards:::.rename_last_group_as_variable(data, by = "ARM", variables = "AESOC")
}
\keyword{internal}
