% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.rnm_grp_vars}
\alias{.rnm_grp_vars}
\title{Rename Group Variables}
\usage{
.rnm_grp_vars(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame}
}
\value{
data frame
}
\description{
This function combines each pair of \code{group} and \code{group_level} columns into a
single column. The \code{group_level} column is renamed according to the value of
the \code{group} column.
}
\examples{
data = data.frame(group1 = "A", x = "B", group2 = "C", y = "D")

cards:::.rnm_grp_vars(data)
}
\keyword{internal}
