% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_ard_columns.R
\name{.shift_column_pair}
\alias{.shift_column_pair}
\title{Shift column pair}
\usage{
.shift_column_pair(x, col_pair)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame}

\item{col_pair}{(\code{character})\cr
character vector containing the column names for variables (first element) and their corresponding levels (second element)}
}
\value{
a tibble
}
\description{
This function ingests an ARD object and coalesces/renames a given pair of columns (variable and levels)
}
\examples{
ard_categorical(ADSL, by = "ARM", variables = "AGEGR1") |>
  cards:::.shift_column_pair(col_pair = c("group1", "group1_level"))

}
\keyword{internal}
