% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical.R
\name{.unnest_table_object}
\alias{.unnest_table_object}
\title{Unnest \code{table()} to ARD Structure}
\usage{
.unnest_table_object(df_result, data)
}
\arguments{
\item{df_result}{an ARD data frame}

\item{data}{the data frame used to construct the ARD data frame. This is used
to extract the variable levels and add to the unnested data frame.}
}
\value{
an ARD data frame
}
\description{
This function takes the 'statistic' returned from the \code{table()} function,
and unnests it to the ARD format including the \code{"variable_level"} column
and the scalar statistics (i.e. \code{"n"} and \code{"p"}).
}
\examples{
ard_continuous(
  mtcars,
  variables = cyl,
  statistics = cyl ~ list(table = function(x) table(x))
) |>
  cards:::.unnest_table_object(data = mtcars)
}
\keyword{internal}
