% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.update_with_fmt_fn}
\alias{.update_with_fmt_fn}
\title{Update Formatting Functions}
\usage{
.update_with_fmt_fn(x, fmt_fn)
}
\arguments{
\item{x}{a ARD object of class 'cards'}

\item{fmt_fn}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character))}.

See the \code{\link{selecting_syntax}} help file for details.}
}
\value{
a ARD object of class 'cards'
}
\description{
Updates the default formatting functions with those passed in this function.
}
\keyword{internal}
