% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_hand_phe}
\alias{eval_hand_phe}
\title{Evaluate a poker hand using PH Evaluator}
\usage{
eval_hand_phe(hand)
}
\arguments{
\item{hand}{a hand of cards (an integer vector of length 5).}
}
\value{
string hand rank.
}
\description{
Evaluate the rank category of a five card poker hand using
\href{https://github.com/HenryRLee/PokerHandEvaluator}{PH Evaluator}.
}
\details{
Implemented following
\href{https://cran.r-project.org/web/packages/Rcpp/vignettes/Rcpp-libraries.pdf}{Rcpp-libraries},
PH Evaluator
\href{https://github.com/HenryRLee/PokerHandEvaluator/blob/master/cpp/examples/cpp_example.cc}{\code{cpp_example.cc}},
and \href{https://github.com/eddelbuettel/rcppannoy}{RcppAnnoy}, with help from
\href{https://teuder.github.io/rcpp4everyone_en/}{Rcpp for everyone}, the rxylib
\href{https://github.com/R-Lum/rxylib/blob/master/src/Makevars}{\code{Makefile}} and
\href{https://stackoverflow.com/questions/43597632/understanding-the-contents-of-the-makevars-file-in-r-macros-variables-r-ma}{Stack Overflow}.

\code{eval_hand_phe} returns one of the rank categories "Straight Flush", "Four of a Kind",
"Full House", "Flush", "Straight", "Three of a Kind", "Two Pair", "One Pair", or "High Card".

"Royal Flush" and "Jacks or Better" are not currently supported.
}
