% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashcard_pdf.R
\name{flashcard_pdf}
\alias{flashcard_pdf}
\title{flashcards}
\usage{
flashcard_pdf(ggplot_object, pdf_name = "student_cards.pdf",
  front_back = TRUE)
}
\arguments{
\item{ggplot_object}{the object that is output from the ggplot_faces function}

\item{pdf_name}{a character string with the name of the file to be saved.  Will be saved to working directory.}

\item{front_back}{states wether to make the pdf output work for a front back printer or to create two pdfs for a basic printer.  Defaults to `TRUE` for a front and back printer.}
}
\description{
flashcards
}
