% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_ard.R
\name{flatten_ard}
\alias{flatten_ard}
\title{Flatten ARD}
\usage{
flatten_ard(x)
}
\arguments{
\item{x}{an ARD object}
}
\value{
data frame
}
\description{
This function ingests an ARD object and flattens it for a readable print of the data.

THIS FUNCTION WILL NEED UPDATES AS SOME FLAVORS OF ARD MAY BECOME MORE COMPLEX.
List columns are converted to character vectors. If a list column contains
elements are cannot easily be coerced to a character vector, the function
will fail.
}
\examples{
ard_categorical(mtcars, by = cyl, include = c(am, gear)) |>
  flatten_ard()
}
