% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ard_stat_string.R
\name{get_ard_stat_string}
\alias{get_ard_stat_string}
\title{ARD Statistics as String}
\usage{
get_ard_stat_string(x, ..., pattern, missing = "")
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}

\item{...}{(\code{\link[rlang:dyn-dots]{dynamic-dots}})\cr
arguments indicating the rows of the ARD to subset.
For example, to return only rows where the column \code{"AGEGR1"} is \code{"65-80"},
pass \code{AGEGR1 \%in\% "65-80"}.
After subsetting, each statistic named in \code{pattern} must appear in exactly one row.}

\item{pattern}{(\code{string})\cr
a string with the statistics to include in glue syntax,
e.g. \code{"{n} / {N} ({p}\%)"}.
The names in the curly brackets must be values of the \code{"stat_name"} column.}

\item{missing}{(\code{string})\cr
string returned when all statistics in \code{pattern} are missing.
Default is \code{""}.}
}
\value{
a string
}
\description{
Returns formatted statistics from an ARD, combined into a single string.
The formatted statistics (the \code{"stat_fmt"} column, see \code{\link[=apply_fmt_fun]{apply_fmt_fun()}})
are inserted into \code{pattern} using glue syntax,
where each \code{{stat_name}} is replaced by the corresponding formatted statistic.
}
\examples{
ard <- ard_tabulate(ADSL, by = "ARM", variables = "AGEGR1")

get_ard_stat_string(
  ard,
  group1_level \%in\% "Placebo",
  variable_level \%in\% "65-80",
  pattern = "{n} / {N} ({p}\%)"
)
}
