% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-checks.R
\name{get_cli_abort_call}
\alias{get_cli_abort_call}
\title{Get Call Environment for \code{\link[cli:cli_abort]{cli::cli_abort()}}}
\usage{
get_cli_abort_call()
}
\description{
Get Call Environment for \code{\link[cli:cli_abort]{cli::cli_abort()}}
}
\seealso{
\code{\link[=set_cli_abort_call]{set_cli_abort_call()}}
}
\keyword{internal}
