% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_faces.R
\name{ggplot_faces}
\alias{ggplot_faces}
\title{@title ggplot faces
	@param faces This variable uses the all_faces object in the list ouput from the jpeg_roll function.
	@param faces_names This variable uses the all_names object in the list ouput from the jpeg_roll function.}
\usage{
ggplot_faces(faces, faces_names)
}
\description{
@title ggplot faces
	@param faces This variable uses the all_faces object in the list ouput from the jpeg_roll function.
	@param faces_names This variable uses the all_names object in the list ouput from the jpeg_roll function.
}
