% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_grid.R
\name{gridServer}
\alias{gridServer}
\title{Metric Grid module's server logic}
\usage{
gridServer(id, cards)
}
\arguments{
\item{id}{a module id name}

\item{metrics}{placeholder}
}
\description{
Metric Grid module's server logic
}
\keyword{internal}
