% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_header.R
\name{header_plan_simple}
\alias{header_plan_simple}
\title{Column Header Plan}
\usage{
header_plan_simple(ard, header = "{strata}  \\nN = {n}")
}
\arguments{
\item{ard}{a ARD object}

\item{header}{a string specifying the form of a column header.
Default is \code{"{strata}  \\nN = {n}"}}
}
\value{
named list
}
\description{
Create a named list column header definitions for statistic columns
}
\examples{
ard_categorical(mtcars, include = cyl) |>
  header_plan_simple(header = "**{strata} Cylinders**  \nN = {n} ({p}\%)")
}
