% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg_roll.R
\name{jpeg_roll}
\alias{jpeg_roll}
\title{Pictorial Class Roll to by student}
\usage{
jpeg_roll(image_files, plot_color = TRUE)
}
\arguments{
\item{image_files}{the image files that are to be converted}

\item{plot_color}{plot the color shading values for the pixels while data is being digested in the function. Defaults to TRUE}
}
\description{
Pictorial Class Roll to by student
}
