% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{json}
\alias{json}
\alias{ard_as_json}
\alias{write_ard_as_json}
\alias{as_nested_list}
\title{ARD as JSON}
\usage{
ard_as_json(x, ...)

write_ard_as_json(x, path, ...)

as_nested_list(x)
}
\arguments{
\item{x}{an ARD}

\item{...}{additional conversion arguments, see also \code{jsonlite::toJSON()}}

\item{path}{file on disk}
}
\value{
a JSON object
}
\description{
Convert ARDs to JSON or write to disk as JSON
}
\examples{
ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, variables = c("mpg", "hp")),
    ard_continuous(mtcars, by = c(cyl, gear), variables = c("mpg", "hp"))
  )

ard_as_json(ard)
}
