% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fmt_fn.R
\name{label_cards}
\alias{label_cards}
\title{Generate Formatting Function}
\usage{
label_cards(digits = 1, scale = 1, width = NULL)
}
\arguments{
\item{digits}{(\code{integer})\cr
a non-negative integer specifying the number of decimal places
round statistics to}

\item{scale}{(\code{numeric})\cr
a scalar real number. Before rounding, the input will be scaled by
this quantity}

\item{width}{(\code{integer})\cr
a non-negative integer specifying the minimum width of the
returned formatted values}
}
\value{
a function
}
\description{
Returns a function with the requested rounding and scaling schema.
}
\examples{
label_cards(2)(pi)
label_cards(1, scale = 100)(pi)
label_cards(2, width = 5)(pi)
}
