% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leading_na}
\alias{leading_na}
\title{Leading NA}
\usage{
leading_na(x)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if
there are leading NA, marking the leading NAs as TRUE and
everything else as FALSE.
}
\details{
Code from R-Bloggers \href{https://www.r-bloggers.com/2016/08/rcpp-and-roxygen2/}{Rcpp and Roxygen2}.

Installed with help from \code{\link[usethis:use_rcpp]{usethis::use_rcpp()}} and roxygen2 instructions on
\href{https://roxygen2.r-lib.org/articles/roxygen2.html#rcpp}{Rcpp}.

Steps to install:
\enumerate{
\item Create \code{src/leading_na.cpp} (this file)
\item Run \code{\link[usethis:use_rcpp]{usethis::use_rcpp()}}, add \verb{@importFrom Rcpp sourceCpp} and
\verb{@useDynLib cards, .registration = TRUE} to \code{package.R} as directed
\item Run \code{\link[desc:desc_normalize]{desc::desc_normalize()}} and \code{\link[devtools:document]{devtools::document()}}
}
}
