% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_2image.R
\name{pdf_2image}
\alias{pdf_2image}
\title{PDF to image conversion (imagemagick)}
\usage{
pdf_2image(file = file.choose(), image_type = "jpg", path_out = NULL)
}
\arguments{
\item{magick_path}{the full character path to the imagemagick convert command.  For example, "C:\\Program Files\\ImageMagick-6.9.2-Q16\\convert.exe" is the default.
@param image_type the type of image format to use in the conversion. Will only do one conversion time per function call.  Input as character and 'jpg' is the default.}

\item{files}{the files to be converted.  Input as a list of character strings.  The working directory will be pre appended to the file names.
@examples pdf_2image(files=file.choose())}
}
\description{
PDF to image conversion (imagemagick)
}
