% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_header.R
\name{plan_header_simple}
\alias{plan_header_simple}
\title{Column Header Plan}
\usage{
plan_header_simple(ard, header = "{strata}  \\nN = {n}")
}
\arguments{
\item{ard}{a ARD object}

\item{header}{a string specifying the form of a column header.
Default is \code{"{strata}  \\nN = {n}"}}
}
\value{
named list
}
\description{
Create a named list column header definitions for statistic columns
}
\examples{
# TODO: The ARD creation code can by simplified after the categorical
#       ARD accepts no-by variable specifications
ard_header <-
  mtcars |>
  dplyr::mutate(..one.. = 1L) |>
  ard_categorical(by = ..one..,  include = cyl) |>
  dplyr::select(-starts_with("strata"))

ard_header |>
  plan_header_simple(header = "**{strata}** Cylinders  \nN = {n}  ({p}\%)")
}
