% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_simple.R
\name{plan_simple}
\alias{plan_simple}
\alias{plan_simple_continuous}
\alias{plan_simple_categorical}
\title{Simple ARD Plans}
\usage{
plan_simple_continuous(
  ard,
  statistics = c("{N}", "{mean} ({sd})", "{min}, {max}")
)

plan_simple_categorical(ard, statistics = c("{n}", "({p})\%"))
}
\arguments{
\item{ard}{a single ARD object for one variable}

\item{statistics}{a character vector of statistics to include in
resulting table. Default is \code{c("{N}", "{mean} ({sd})",  "{min}, {max}")}}
}
\value{
data frame
}
\description{
Plans to convert simple ARDs to formatted data frames.
These data frames will eventually be stacked into a final table.
}
\examples{
ard_continuous(mtcars, by = cyl, include = mpg) |>
  plan_simple_continuous()
}
