% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcard.R
\name{postcard}
\alias{postcard}
\title{Create a postcard}
\usage{
postcard(
  width,
  height,
  text,
  text_width,
  text_x,
  text_y,
  family,
  background_colour,
  text_colour
)
}
\arguments{
\item{width}{Width}

\item{height}{Height}

\item{text}{Text}

\item{text_width}{Text Width}

\item{text_x}{Text X}

\item{text_y}{Text Y}

\item{family}{Font family}

\item{background_colour}{Background Colour}

\item{text_colour}{Text Colour}
}
\value{
A ggplot2 object
}
\description{
Create a postcard
}
