% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cards}
\alias{print.cards}
\title{Print Methods}
\usage{
\method{print}{cards}(x, ...)
}
\arguments{
\item{x}{a cards table}

\item{...}{Not used}
}
\value{
a gt table
}
\description{
Print Methods
}
\examples{
# TODO: add example
}
