% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{print_hand}
\alias{print_hand}
\title{Print a hand of cards}
\usage{
print_hand(hand, collapse = TRUE)
}
\arguments{
\item{hand}{a hand of cards.}

\item{collapse}{if \code{TRUE}, collapse the character vector.}
}
\description{
Returns a character vector of cards, collapsed by default.
}
\examples{
(hand <- deal_hand(new_deck()))
print_hand(hand)
}
