% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{all_of}
\alias{any_of}
\alias{everything}
\alias{where}
\alias{last_col}
\alias{one_of}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:\%>\%]{\%>\%}}, \code{\link[dplyr:all_of]{all_of}}, \code{\link[dplyr:any_of]{any_of}}, \code{\link[dplyr:contains]{contains}}, \code{\link[dplyr:ends_with]{ends_with}}, \code{\link[dplyr:everything]{everything}}, \code{\link[dplyr:last_col]{last_col}}, \code{\link[dplyr:matches]{matches}}, \code{\link[dplyr:num_range]{num_range}}, \code{\link[dplyr:one_of]{one_of}}, \code{\link[dplyr:starts_with]{starts_with}}, \code{\link[dplyr:where]{where}}}
}}

