% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_data.R
\name{register_data}
\alias{register_data}
\alias{add_registered_data_column}
\alias{print.cards_data_desc}
\title{Register Data Frame}
\usage{
register_data(
  data,
  filter_expr = TRUE,
  data_name = rlang::caller_arg(data),
  filter_label = rlang::expr_deparse(filter_expr, width = Inf)
)

add_registered_data_column(x, data)

\method{print}{cards_data_desc}(x)
}
\arguments{
\item{data}{a data frame, e.g. \code{ADSL}.}

\item{filter_expr}{an expression that will be applied to the data frame passed in \code{data}.}

\item{data_name}{string of the data frame name. Default is \code{rlang::caller_arg(data)}.}

\item{filter_label}{string describing the filter. The default is the string version
filter expression.}
}
\value{
data frame
}
\description{
\itemize{
\item \code{register_data()}: Register a data frame
\item \code{add_registered_data_column()}: Add the data registration information to the ARD
}
}
