% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Rounding of Numbers}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\value{
numeric vector
}
\description{
Rounds the values in its first argument to the specified number of
decimal places (default 0). Importantly, \code{round2()} \strong{does not} use Base R's
"round to even" default. Standard rounding methods are implemented, for example,
\code{round2(0.5) = 1}.
}
\details{
Function inspired by \code{janitor::round_half_up()}.
}
\examples{
x <- 0:4 / 2
round2(x)

# compare results to Base R
round(x)
}
