% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-checks.R
\name{set_cli_abort_call}
\alias{set_cli_abort_call}
\title{Set Call Environment for \code{\link[cli:cli_abort]{cli::cli_abort()}}}
\usage{
set_cli_abort_call(env = caller_env())
}
\arguments{
\item{env}{(\code{enviroment})\cr
call environment used as the \code{call} parameter in \code{\link[cli:cli_abort]{cli::cli_abort()}} for package checks}
}
\description{
Set a call environment to be used as the \code{call} parameter in \code{\link[cli:cli_abort]{cli::cli_abort()}} for package checks. This function
is used to ensure that the correct user-facing function is reported for errors generated by internal checks that
use \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\seealso{
\code{\link[=get_cli_abort_call]{get_cli_abort_call()}}
}
\keyword{internal}
