% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ard_order.R
\name{tidy_ard_column_order}
\alias{tidy_ard_column_order}
\title{Standard Column Order of ARD}
\usage{
tidy_ard_column_order(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a ARD data frame of class 'card'}
}
\value{
a data frame
}
\description{
Relocate all columns in ARD to the standard order.
}
\examples{
ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, variables = "mpg"),
    ard_continuous(mtcars, variables = "mpg", by = "cyl")
  )
names(ard)

tidy_ard_column_order(ard)
}
