% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ard_order.R
\name{tidy_ard_row_order}
\alias{tidy_ard_row_order}
\title{Order rows of ARD}
\usage{
tidy_ard_row_order(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a ARD data frame of class 'card'}
}
\value{
a data frame
}
\description{
Order rows of ARD according to variables, groups, and strata, retaining the
order of input ARDs
}
\examples{
ard <-
  bind_ard(
    ard_continuous(mtcars, variables = c("mpg","disp"), by = "cyl"),
    ard_ttest(mtcars, variable = "mpg", by = "cyl"),
    ard_ttest(mtcars, variable = "disp", by = "cyl")
  )
tidy_ard_row_order(ard)
}
