% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-helpers.R
\name{ard-helpers}
\alias{ard-helpers}
\alias{get_ard_statistics}
\alias{filter_ard}
\title{ARD Helpers}
\usage{
get_ard_statistics(x, ...)

filter_ard(x, ...)
}
\arguments{
\item{x}{an ARD data frame of class 'card'}

\item{...}{optional named arguments indicating rows to subset of the ARD.
For example, to return only rows where the column \code{"AGEGR1"} is \code{"65-80"},
pass \code{AGEGR1 = "65-80"}.}
}
\value{
a transformed ARD
}
\description{
Helper functions for working with ARD data frames
}
\examples{
ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")

filter_ard(ard, group1_level = "Placebo", variable_level = "65-80") |>
  flatten_ard()

get_ard_statistics(ard, group1_level = "Placebo", variable_level = "65-80")
}
