% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard-helpers.R
\name{ard-helpers}
\alias{ard-helpers}
\alias{get_ard_statistics}
\alias{bind_ard}
\title{ARD Helpers}
\usage{
get_ard_statistics(
  x,
  ...,
  .column = "statistic",
  .attributes = c("stat_label", "statistic_fmt_fn", "warning", "error")
)

bind_ard(..., .update = FALSE)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}

\item{...}{\itemize{
\item \code{get_ard_statistics()}: optional named arguments indicating rows to subset of the ARD.
For example, to return only rows where the column \code{"AGEGR1"} is \code{"65-80"},
pass \code{AGEGR1 \%in\% "65-80"}.
\item \code{bind_ard()}: ARDs to combine. Each argument can either be an ARD,
or a list of ARDs. Columns are matched by name, and any missing
columns will be filled with \code{NA.}
}}

\item{.column}{(\code{string})\cr
string indicating the column that will be returned in the list.
Default is \code{"statistic"}}

\item{.attributes}{(\code{character})\cr
character vector of column names that will be returned
in the list as attributes.
Default is \code{c("stat_label", "statistic_fmt_fn", "warning", "error")}}

\item{.update}{(\code{logical} scalar)\cr
logical indicating whether to update duplicate ARD statistics.
Default is \code{FALSE}. If a statistic type is repeated and \code{.update=TRUE},
the more recently added statistics will be retained, and the others omitted.}
}
\value{
a transformed ARD
}
\description{
Helper functions for working with ARD data frames.
}
\details{
\itemize{
\item \code{get_ard_statistics()}: This function returns the statistics from an ARD
as a named list. The statistic label, formatting instruction, and any error
or warning messages are returned as attributes.
\item \code{bind_ard()}: A wrapper for \code{dplyr::bind_rows()} with additional checks
for duplicated statistics
}
}
\examples{
ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")

get_ard_statistics(ard, group1_level \%in\% "Placebo", variable_level \%in\% "65-80")

bind_ard(ard, ard, .update = TRUE)
}
