% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_attributes.R
\name{ard_attributes}
\alias{ard_attributes}
\alias{ard_attributes.data.frame}
\alias{ard_attributes.default}
\title{Add variable attributes to an ARD object}
\usage{
ard_attributes(data, ...)

\method{ard_attributes}{data.frame}(data, variables = everything(), label = NULL, ...)

\method{ard_attributes}{default}(data, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{...}{Arguments passed to other methods}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables to include}

\item{label}{(named \code{list})\cr
named list of variable labels, e.g. \code{list(cyl = "No. Cylinders")}.
Default is \code{NULL}}
}
\value{
an ARD data frame of class 'card'
}
\description{
S3 generic providing methods for the following classes:
\itemize{
\item \code{data.frame}: Adds variable attributes to an ARD data frame.
}
}
\examples{
df <- dplyr::tibble(var1 = letters, var2 = LETTERS)
attr(df$var1, "label") <- "Lowercase Letters"

ard_attributes(df, variables = everything())

}
