% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_attributes.R
\name{ard_attributes}
\alias{ard_attributes}
\alias{ard_attributes.data.frame}
\alias{ard_attributes.default}
\title{ARD Attributes}
\usage{
ard_attributes(x, ...)

\method{ard_attributes}{data.frame}(x, variables = everything(), label = NULL, ...)

\method{ard_attributes}{default}(x, ...)
}
\arguments{
\item{x}{The R object you want to add variable attributes to}

\item{...}{Arguments passed to other methods}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables to include}

\item{label}{(named \code{list})\cr
named list of variable labels, e.g. \code{list(cyl = "No. Cylinders")}.
Default is \code{NULL}}
}
\value{
an ARD data frame of class 'card'
}
\description{
S3 generic providing methods for the following classes:
\itemize{
\item \code{data.frame}: Adds variable attributes to an ARD data frame.
}

Add variable attributes to an ARD data frame.
\itemize{
\item The \code{label} attribute will be added for all columns, and when no label
is specified and no label has been set for a column using the \verb{label=} argument,
the column name will be placed in the label statistic.
\item The \code{class} attribute will also be returned for all columns.
\item Any other attribute returned by \code{attributes()} will also be added, e.g. factor levels.
}
}
\examples{
df <- dplyr::tibble(var1 = letters, var2 = LETTERS)
attr(df$var1, "label") <- "Lowercase Letters"

ard_attributes(df, variables = everything())
}
