% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical.R
\name{ard_categorical}
\alias{ard_categorical}
\title{Categorical ARD Statistics}
\usage{
ard_categorical(
  data,
  variables,
  by = NULL,
  strata = NULL,
  statistics = everything() ~ categorical_variable_summary_fns(),
  denominator = NULL,
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)
}
\arguments{
\item{data}{a data frame}

\item{variables}{columns to include in summaries. Default is \code{everything()}.}

\item{by, strata}{columns to by/stratified by for tabulation.
Arguments are similar, but with an important distinction:

\code{by}: results are tabulated by \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.

\code{strata}: results are tabulated by \strong{all \emph{observed} combinations} of the
columns specified.

Arguments may be used in conjunction with one another.}

\item{statistics}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) mean(x)))}.

The value assigned to each variable must also be a named list, where the names
are used to reference a function and the element is the function object.
Typically, this function will return a scalar statistic, but a function that
returns a named list of results is also acceptable, e.g.
\code{list(conf.low = -1, conf.high = 1)}. However, when errors occur, the messaging
will be less clear in this setting.

See the \code{\link{selecting_syntax}} help file for details.}

\item{denominator}{Specify this \emph{optional} argument to change the denominator,
e.g. the \code{"N"} statistic. Default is \code{NULL}. See below for details.}

\item{fmt_fn}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round5(x, digits = 2) |> as.character))}.

See the \code{\link{selecting_syntax}} help file for details.}

\item{stat_labels}{a named list, a list of formulas, or a single formula where
the list element is either a named list or a list of formulas defining the
statistic labels, e.g. \code{everything() ~ list(n = "n", p = "pct")} or
\code{everything() ~ list(n ~ "n", p ~ "pct")}.}
}
\value{
a data frame
}
\description{
Compute Analysis Results Data (ARD) for categorical summary statistics.
}
\section{Denominators}{

By default, the \code{ard_categorical()} function returns the statistics \code{"n"} and \code{"N"},
where little \code{"n"} are the counts for the variable levels, and \code{"N"} is
the number of non-missing observations. The default calculation for the
percentage is merely \code{p = n/N}.

However, it is sometimes necessary to provide a different \code{"N"} to use
as the denominator in this calculation. For example, in a calculation
of the rates of various observed adverse events, you may need to update the
denominator to the number of enrolled subjects.

In such cases, use the \code{denominator} argument to specify a new definition
of \code{"N"}, and subsequently \code{"p"}.
The argument expects a data frame, and the data frame must include the columns
specified in \code{ard_categorical(by=)} (strata columns are not considered).
The updated \code{N} and \code{length} elements will be updated to be calculated as
the number of rows in each combination of the \code{by} variables.

Take an example where we need to update the denominator to be subjects enrolled
in a trial, e.g. tabulating the number of AEs that occurred within an SOC
where some subjects may not have experienced an AE and would not be represented
in the ADAE data set. All patients appear in ADSL, however.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ard_categorical(
  data =
    ADAE |>
      dplyr::left_join(ADSL[c("USUBJID", "ARM")], by = "USUBJID") |>
      dplyr::filter(AOCCSFL \%in\% "Y"),
  by = "ARM",
  variables = "AESOC",
  denominator = ADSL
) |>
  flatten_ard()
}\if{html}{\out{</div>}}
}

\examples{
ard_categorical(ADSL, by = "ARM", variables = "AGEGR1") |>
  flatten_ard()
}
