% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_comparison.R
\name{ard_comparison}
\alias{ard_comparison}
\alias{ard_ttest}
\alias{ard_wilcoxtest}
\alias{ard_chisqtest}
\alias{ard_fishertest}
\title{Comparison ARD Statistics}
\usage{
ard_ttest(data, by, variable, ...)

ard_wilcoxtest(data, by, variable, ...)

ard_chisqtest(data, by, variable, ...)

ard_fishertest(data, by, variable, ...)
}
\arguments{
\item{data}{a data frame}

\item{by}{character column name to compare by}

\item{variable}{character column name to be compared}

\item{...}{arguments passed to method.}
}
\value{
data frame
}
\description{
ARD functions for comparing values between groups.

\code{ard_ttest()} -> \code{t.test(data[[variable]] ~ data[[by]], ...)}
\code{ard_wilcoxtest()} -> \code{wilcox.test(data[[variable]] ~ data[[by]], ...)}
\code{ard_chisqtest()} -> \code{chisq.test(x = data[[variable]], y = data[[by]], ...)}
\code{ard_fishertest()} -> \code{fisher.test(x = data[[variable]], y = data[[by]], ...)}
}
\examples{
ADSL |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  ard_ttest(by = "ARM", variable = "AGE") |>
  flatten_ard()

ADSL |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  ard_wilcoxtest(by = "ARM", variable = "AGE") |>
  flatten_ard()

ADSL |>
  ard_chisqtest(by = "ARM", variable = "AGEGR1") |>
  flatten_ard()

ADSL[1:30,] |>
  ard_fishertest(by = "ARM", variable = "AGEGR1") |>
  flatten_ard()
}
