% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_comparison.R
\name{ard_comparison}
\alias{ard_comparison}
\alias{ard_ttest}
\alias{ard_wilcoxtest}
\title{Comparison ARD Statistics}
\usage{
ard_ttest(data, by, variable, ...)

ard_wilcoxtest(data, by, variable, ...)
}
\arguments{
\item{data}{a data frame}

\item{by}{charcter column name to compare by}

\item{variable}{charadter column name to be compared}

\item{...}{arguments passed to method.}
}
\value{
data frame
}
\description{
Comparison ARD Statistics
}
\examples{
ard_ttest(data = mtcars, by = "am", variable = "hp")
ard_wilcoxtest(data = mtcars, by = "am", variable = "hp")
}
