% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{ard_continuous}
\alias{ard_continuous}
\title{Continuous ARD Statistics}
\usage{
ard_continuous(
  data,
  variables,
  by = NULL,
  strata = NULL,
  statistics = everything() ~ continuous_variable_summary_fns(),
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)
}
\arguments{
\item{data}{a data frame}

\item{variables}{columns to include in summaries.}

\item{by, strata}{columns to by/stratified by for summary statistic
calculation. Arguments are similar, but with an important distinction:

\code{by}: results are calculated by \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.

\code{strata}: results are calculated by \strong{all \emph{observed} combinations} of the
columns specified.

Arguments may be used in conjunction with one another.}

\item{statistics}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) mean(x)))}.

The value assigned to each variable must also be a named list, where the names
are used to reference a function and the element is the function object.
Typically, this function will return a scalar statistic, but a function that
returns a named list of results is also acceptable, e.g.
\code{list(conf.low = -1, conf.high = 1)}. However, when errors occur, the messaging
will be less clear in this setting.

See the \code{\link{selecting_syntax}} help file for details.}

\item{fmt_fn}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character))}.

See the \code{\link{selecting_syntax}} help file for details.}

\item{stat_labels}{a named list, a list of formulas, or a single formula where
the list element is either a named list or a list of formulas defining the
statistic labels, e.g. \code{everything() ~ list(mean = "Mean", sd = "SD")} or
\code{everything() ~ list(mean ~ "Mean", sd ~ "SD")}.}
}
\value{
a data frame
}
\description{
Compute Analysis Results Data (ARD) for simple continuous summary statistics.
}
\examples{
ard_continuous(ADSL, by = "ARM", variables = "AGE") |>
  flatten_ard()
}
