% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_crosstab.R
\name{ard_crosstab}
\alias{ard_crosstab}
\title{ARD Crosstab}
\usage{
ard_crosstab(data, variables, by)
}
\arguments{
\item{data}{a data frame}

\item{variables}{columns to include in summaries. Default is \code{everything()}.}

\item{by}{columns to compute statistics by.}
}
\value{
a data frame
}
\description{
Similar to \code{ard_categorical()} except the percentages and N statistics are
based on the entire data set, rather than stratified columns.
}
\examples{
ard_crosstab(ADSL, variables = "AGEGR1", by = "ARM")
}
