% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_hierarchical.R
\name{ard_hierarchical}
\alias{ard_hierarchical}
\alias{ard_hierarchical_count}
\title{Title}
\usage{
ard_hierarchical(
  data,
  variables,
  by = NULL,
  statistics = everything() ~ categorical_variable_summary_fns(),
  denominator = NULL,
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)

ard_hierarchical_count(
  data,
  variables,
  by = NULL,
  fmt_fn = NULL,
  stat_labels = everything() ~ default_stat_labels()
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables to perform the nested/hierarchical tabulations within}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables to perform tabulations by. All combinations of the variables
specified here appear in results. Default is \code{NULL}}

\item{statistics}{(\code{\link[=selecting_syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) mean(x)))}.

The value assigned to each variable must also be a named list, where the names
are used to reference a function and the element is the function object.
Typically, this function will return a scalar statistic, but a function that
returns a named list of results is also acceptable, e.g.
\code{list(conf.low = -1, conf.high = 1)}. However, when errors occur, the messaging
will be less clear in this setting.}

\item{denominator}{(\code{data.frame})\cr
Specify this \emph{optional} argument to change the denominator,
e.g. the \code{"N"} statistic. Default is \code{NULL}. See below for details.}

\item{fmt_fn}{(\code{\link[=selecting_syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character))}.}

\item{stat_labels}{(\code{\link[=selecting_syntax]{formula-list-selector}})\cr
a named list, a list of formulas, or a single formula where
the list element is either a named list or a list of formulas defining the
statistic labels, e.g. \code{everything() ~ list(n = "n", p = "pct")} or
\code{everything() ~ list(n ~ "n", p ~ "pct")}.}
}
\value{
an ARD data frame
}
\description{
Title
}
\examples{
ard_hierarchical(
  data = ADAE,
  variables = c(AESOC, AETERM),
  by = c(TRTA, AESEV),
  denominator = ADSL |> dplyr::rename(TRTA = ARM)
) |>
  flatten_ard()

ard_hierarchical_count(
  data = ADAE,
  variables = c(AESOC, AETERM),
  by = TRTA
) |>
  flatten_ard()
}
