% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_missing.R
\name{ard_missing}
\alias{ard_missing}
\title{Missing ARD Statistics}
\usage{
ard_missing(
  data,
  variables,
  by = NULL,
  statistics = everything() ~ missing_variable_summary_fns(),
  fmt_fn = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{variables}{columns to include in summaries.}

\item{by}{results are tabulated by \strong{all combinations} of the columns specified}

\item{statistics}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) mean(x)))}.

The value assigned to each variable must also be a named list, where the names
are used to reference a function and the element is the function object.
Typically, this function will return a scalar statistic, but a function that
returns a named list of results is also acceptable, e.g.
\code{list(conf.low = -1, conf.high = 1)}. However, when errors occur, the messaging
will be less clear in this setting.

See the \code{\link{selecting_syntax}} help file for details.}

\item{fmt_fn}{a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character))}.

See the \code{\link{selecting_syntax}} help file for details.}
}
\value{
a data frame
}
\description{
Compute Analysis Results Data (ARD) for statistics about data missingness.
}
\examples{
ard_missing(ADSL, by = "ARM", variables = "AGE") |>
  flatten_ard()
}
