% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_simple.R
\name{ard_simple}
\alias{ard_simple}
\alias{ard_continuous}
\alias{ard_categorical}
\title{Simple ARD Statistics}
\usage{
ard_continuous(
  data,
  by = dplyr::group_vars(data),
  statistics = NULL,
  include = everything()
)

ard_categorical(data, by = dplyr::group_vars(data), include = everything())
}
\arguments{
\item{data}{a data frame}

\item{by}{columns to compute statistics by. Default are the columns
returned by \code{dplyr::group_vars(data)}.}

\item{statistics}{a named list of functions that return a summary statistic,
e.g. \verb{list(mpg = list(mean = \\(x) mean(x, na.rm = TRUE)))}}

\item{include}{columns to include in summaries. Default is \code{everything()}.}
}
\value{
a data frame
}
\description{
Compute Analysis Results Data (ARD) for simple summary statistics from
continuous and categorical data.
}
\examples{
ard_continuous(mtcars, by = cyl, include = c(mpg, hp))
ard_categorical(mtcars, by = cyl, include = c(am, gear))
}
