% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_stack_hierarchical.R
\name{ard_stack_hierarchical}
\alias{ard_stack_hierarchical}
\title{Stacked Hierarchical ARD Statistics}
\usage{
ard_stack_hierarchical(
  data,
  variables,
  by = dplyr::group_vars(data),
  id = NULL,
  denominator = NULL,
  include = everything(),
  overall = FALSE,
  overall_row = FALSE,
  attributes = FALSE,
  total_n = FALSE,
  shuffle = FALSE
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Specifies the nested/hierarchical structure of the data.
The variables that are specified here and in the \code{include} argument
will have summary statistics calculated.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables to perform tabulations by. All combinations of the variables
specified here appear in results. Default is \code{dplyr::group_vars(data)}.}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
an optional argument used to subset \code{data} to obtain the correct percentages.
When specified, \code{ard_hierarchical()} is used to calculate event rates. When
not specified, counts are returned via \code{ard_hierarchical_count()}.}

\item{denominator}{(\code{data.frame}, \code{integer})\cr
an optional argument that can enhance the output.
\itemize{
\item if \code{total_n=TRUE}, the \code{denominator} argument is used to calculate/return the N
\item the univariate tabulations of the \code{by} variables are calculated with \code{denominator}
\item the \code{denominator} argument must be specified when \code{id} is used to
calculate the event rates. See the \code{ard_hierarchical()} help page for details
}}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Specify the subset a columns indicated in the \code{variables} argument for which
summary statistics will be returned. Default is \code{everything()}.}

\item{overall}{(\code{logical})\cr logical indicating whether overall statistics
should be calculated (i.e. re-run all \verb{ard_*()} calls with \code{by=NULL}).
Default is \code{FALSE}.}

\item{overall_row}{(\code{logical})\cr logical indicating whether overall statistics
should be calculated across the columns listed in the \code{variables} argument.
Default is \code{FALSE}.}

\item{attributes}{(\code{logical})\cr
logical indicating whether to include the results of \code{ard_attributes()} for all
variables represented in the ARD. Default is \code{FALSE}.}

\item{total_n}{(\code{logical})\cr
logical indicating whether to include of \code{ard_total_n(denominator)} in the returned ARD.}

\item{shuffle}{(\code{logical})\cr
logical indicating whether to perform \code{shuffle_ard()} on the final result.
Default is \code{FALSE}.}
}
\value{
an ARD data frame of class 'card'
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Use this function to calculate multiple summaries of nested or hierarchical data
in a single call.

When the \code{id} argument is specified, event rates and counts are calculated via
\code{ard_hierarchical()}; otherwise, counts are calculated using \code{ard_hierarchical_count()}.
See below for details on rate calculations when using the \code{id} argument.
}
\details{
ADD DETAILS ABOUT HOW THE DATA ARE SORTED AND SUBSETTED.
}
\examples{
ard_stack_hierarchical(
  ADAE,
  variables = c(AESOC, AEDECOD),
  by = TRTA,
  denominator = ADSL |> dplyr::rename(TRTA = ARM),
  id = USUBJID
)
}
