% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_stack_hierarchical.R
\name{ard_stack_hierarchical}
\alias{ard_stack_hierarchical}
\title{Title}
\usage{
ard_stack_hierarchical(
  data,
  variables,
  by = dplyr::group_vars(data),
  id = NULL,
  denominator = NULL,
  include = everything(),
  overall = FALSE,
  attributes = FALSE,
  total_n = FALSE,
  shuffle = FALSE
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Specifies the nested/hierarchical structure of the data.
The variables that are specified here and in the \code{include} argument
will have summary statistics calculated.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables to perform tabulations by. All combinations of the variables
specified here appear in results. Default is \code{dplyr::group_vars(data)}.}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
an optional argument used to subset \code{data} to obtain the correct percentages.
When specified, \code{ard_hierarchical()} is used to calculate event rates. When
not specified, counts are returned via \code{ard_hierarchical_count()}.}

\item{denominator}{(\code{data.frame})\cr
Must be specified when the \code{id} argument is used. Pass a data frame here
to indicate what the denominator is meant to be for rate calculations.
See the \code{ard_hierarchical()} help page for details.}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Specify the subset a columns indicated in the \code{variables} argument for which
summary statistics will be returned. Default is \code{everything()}.}

\item{overall}{(\code{logical})\cr logical indicating whether overall statistics
should be calculated (i.e. re-run all \verb{ard_*()} calls with \code{by=NULL}).
Default is \code{FALSE}.}

\item{attributes}{(\code{logical})\cr
logical indicating whether to include the results of \code{ard_attributes()} for all
variables represented in the ARD. Default is \code{FALSE}.}

\item{total_n}{(\code{logical})\cr
logical indicating whether to include of \code{ard_total_n(denominator)} in the returned ARD.}

\item{shuffle}{(\code{logical})\cr
logical indicating whether to perform \code{shuffle_ard()} on the final result.
Default is \code{FALSE}.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Title
}
\examples{
ard_stack_hierarchical(
  ADAE,
  variables = c(AESOC, AEDECOD),
  by = TRTA,
  denominator = ADSL |> dplyr::rename(TRTA = ARM),
  id = USUBJID
)
}
