% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_strata.R
\name{ard_strata}
\alias{ard_strata}
\title{Stratified ARD}
\usage{
ard_strata(data, strata, .ard_fn, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{strata}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to stratify results by.}

\item{.ard_fn}{(\code{function}, \code{formula})\cr
a function or a formula that can be coerced to a function with
\code{rlang::as_function()} (similar to \code{purrr::map(.f)})}

\item{...}{Additional arguments passed on to the \code{.ard_fn} function.}
}
\value{
an ARD data frame of class 'card'
}
\description{
General function for calculating ARD results within subgroups.

While the examples below show use with other functions from the cards package,
this function would primarily be used with the statistical functions in the
cardx functions.
}
\examples{
ard_strata(
  ADSL,
  strata = ARM,
  ~ ard_continuous(.x, variables = AGE)
)

ard_strata(
  ADSL,
  strata = ARM,
  ard_continuous,
  variables = AGE
)
}
