% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_total_n.R
\name{ard_total_n}
\alias{ard_total_n}
\title{ARD Total N}
\usage{
ard_total_n(data)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}
}
\value{
an ARD data frame of class 'card'
}
\description{
Returns the total N for the data frame.
The placeholder variable name returned in the object is \code{"..ard_total_n.."}
}
\examples{
ard_total_n(ADSL)
}
