% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_card.R
\name{as_card}
\alias{as_card}
\title{Data Frame as ARD}
\usage{
as_card(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame}
}
\value{
an ARD data frame of class 'card'
}
\description{
Convert data frames to ARDs of class 'card'.
}
\examples{
data.frame(
  stat_name = c("N", "mean"),
  stat_label = c("N", "Mean"),
  stat = c(10, 0.5)
) |>
  as_card()
}
