% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nested_list.R
\name{as_nested_list}
\alias{as_nested_list}
\title{ARD as Nested List}
\usage{
as_nested_list(x)
}
\arguments{
\item{x}{an ARD}
}
\value{
a nested list
}
\description{
Convert ARDs to Nested Lists
}
\examples{
ard_continuous(mtcars, by = "cyl", variables = c("mpg", "hp")) |>
  as_nested_list()
}
