% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_ard.R
\name{bind_ard}
\alias{bind_ard}
\title{Bind ARDs}
\usage{
bind_ard(..., .update = FALSE)
}
\arguments{
\item{...}{ARDs to combine. Each argument can either be an ARD,
or a list of ARDs. Columns are matched by name, and any missing
columns will be filled with \code{NA.}}

\item{.update}{(\code{logical} scalar)\cr
logical indicating whether to update duplicate ARD statistics.
Default is \code{FALSE}. If a statistic type is repeated and \code{.update=TRUE},
the more recently added statistics will be retained, and the others omitted.}
}
\value{
a transformed ARD
}
\description{
Wrapper for \code{dplyr::bind_rows()} with additional checks
for duplicated statistics.
}
\examples{
ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")

bind_ard(ard, ard, .update = TRUE)
}
