% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Card-class}
\alias{Card-class}
\title{Card}
\description{
A playing card
}
\section{Slots}{

\describe{
\item{\code{suit}}{Is this card a spade, club, diamond, or heart?}

\item{\code{number}}{The card’s value. Can be 2-10, ‘J’, ‘Q’, ‘K’, or ‘A’}
}}

