% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Card}
\alias{Card}
\title{Card constructor function}
\usage{
Card(suit, number, trump = FALSE)
}
\arguments{
\item{suit}{suit of card. Must be one of "spades", "clubs", "diamonds", or "hearts"}

\item{number}{value of card. Must be an integer between 2 and 14, but can be numeric type.
11 through 14 represent Jack through Ace, respectively.}

\item{trump}{set to TRUE if the card is a trump card. default FALSE.}
}
\description{
Card constructor function
}

