% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal-package.R
\docType{package}
\name{cardinal-package}
\alias{cardinal}
\alias{cardinal-package}
\title{cardinal: CDISC Analysis Results Data}
\description{
The package constructs CDISC-compliant Analysis Results Data objects. These objects are used to construct summary tables and visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/cardinal}
  \item \url{https://refactored-bassoon-49o3774.pages.github.io/}
  \item Report bugs at \url{https://github.com/insightsengineering/cardinal/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
