% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards-package.R
\docType{package}
\name{cards-package}
\alias{cards}
\alias{cards-package}
\title{cards: CDISC Analysis Results Data}
\description{
The package constructs CDISC-compliant Analysis Results Data objects. These objects are used (and re-used) to construct summary tables, visualizations, and written reports. The results may also be represented as JSON or YAML objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/cards}
  \item \url{https://refactored-bassoon-49o3774.pages.github.io/}
  \item Report bugs at \url{https://github.com/insightsengineering/cards/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
