% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards-package.R
\docType{package}
\name{cards-package}
\alias{cards}
\alias{cards-package}
\title{cards: Analysis Results Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Construct CDISC (Clinical Data Interchange Standards Consortium) compliant Analysis Results Data objects. These objects are used and re-used to construct summary tables, visualizations, and written reports. The package also exports utilities for working with these objects and creating new Analysis Results Data objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/cards}
  \item \url{https://insightsengineering.github.io/cards/}
  \item Report bugs at \url{https://github.com/insightsengineering/cards/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Becca Krouse \email{becca.z.krouse@gsk.com}
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline Research & Development Limited [copyright holder]
}

}
\keyword{internal}
