% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cards-package}
\alias{cards}
\alias{cards-package}
\title{cards: Card Simulation Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for card game simulations, written in base R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/cards/}
  \item \url{https://github.com/jabenninghoff/cards}
  \item Report bugs at \url{https://github.com/jabenninghoff/cards/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
