% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ard_structure.R
\name{check_ard_structure}
\alias{check_ard_structure}
\title{Check ARD Structure}
\usage{
check_ard_structure(x)
}
\arguments{
\item{x}{an ARD data frame}
}
\value{
an ARD data frame
}
\description{
Function tests the structure and returns notes when object does not
conform to expected structure.
}
\examples{
ard_continuous(ADSL, variables = "AGE") |>
  dplyr::select(-warning, -error) |>
  check_ard_structure()
}
