% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cardinal.R
\name{construct_cardinal}
\alias{construct_cardinal}
\title{Construct Cardinal Table}
\usage{
construct_cardinal(plan_table, plan_header, hide = c("variable", "header_row"))
}
\arguments{
\item{plan_table}{a table plan data frame}

\item{plan_header}{a header plan named list}

\item{hide}{character vector of columns to hide from output}
}
\value{
a cardinal table
}
\description{
Construct a table of class 'cardinal' from the table plan, header plan, and
other table compnents
}
\examples{
# Construct the ARD
table_ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, include = mpg),
    ard_categorical(mtcars, by = cyl, include = am),
    # TODO: The ARD creation code can by simplified after the categorical
    #       ARD accepts no-by variable specifications
    mtcars |>
      dplyr::mutate(..one.. = 1L) |>
      ard_categorical(by = ..one..,  include = cyl) |>
      dplyr::select(-starts_with("strata"))
  )

# convert ARD to a cardinal table
construct_cardinal(
  plan_table =
    dplyr::bind_rows(
      table_ard |> dplyr::filter(variable \%in\% "mpg") |>  plan_table_simple_continuous(),
      table_ard |> dplyr::filter(variable \%in\% "am") |> plan_table_simple_categorical()
    ),
  plan_header =
    table_ard |>
    dplyr::filter(variable \%in\% "cyl") |>
    plan_header_simple(header = "**{strata} Cylinders  \nN = {n}  ({p}\%)**") |>
    modifyList(val = list(label = "**Characteristic**"))
)
}
