% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cardinal.R
\name{convert_cardinal}
\alias{convert_cardinal}
\title{Convert Table to Print Engine Class}
\usage{
convert_cardinal(x, engine = "gt")
}
\arguments{
\item{x}{a cardinal table}

\item{engine}{a string indicating the print engine. Default is \code{"gtß"}}
}
\value{
a table object
}
\description{
Convert a 'cardinal' table to any of the print engine classes.
Goal is to support gt, rtables, flextable, huxtable, kableExtra,
kable, and formatted tibbles
}
\examples{
# Construct the ARD
table_ard <-
  dplyr::bind_rows(
    ard_continuous(mtcars, by = cyl, include = mpg),
    ard_categorical(mtcars, by = cyl, include = am),
    ard_categorical(mtcars,  nclude = cyl)
  )

# convert ARD to a cardinal table
construct_cardinal(
  table_plan =
    dplyr::bind_rows(
      table_ard |> dplyr::filter(variable \%in\% "mpg") |>  table_plan_simple_continuous(),
      table_ard |> dplyr::filter(variable \%in\% "am") |> table_plan_simple_categorical()
    ),
  header_plan =
    table_ard |>
    dplyr::filter(variable \%in\% "cyl") |>
    header_plan_simple(header = "**{strata} Cylinders  \nN = {n}  ({p}\%)**") |>
    modifyList(val = list(label = gt::md("**Characteristic**")))
) |>
  convert_cardinal()
}
