% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_stat_labels.R
\name{default_stat_labels}
\alias{default_stat_labels}
\title{List of stat labels}
\usage{
default_stat_labels()
}
\value{
named list of stat labels
}
\description{
Returns a list of stat labels
}
\examples{
default_stat_labels()
}
