% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{apply_fmt_fn}
\alias{alias_as_fmt_fn}
\alias{update_ard_fmt_fn}
\alias{label_cards}
\title{Deprecated functions}
\usage{
apply_fmt_fn(...)

alias_as_fmt_fn(...)

update_ard_fmt_fn(...)

label_cards(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
