% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.calculate_stats_as_ard}
\alias{.calculate_stats_as_ard}
\title{Calculate Continuous Statistics}
\usage{
.calculate_stats_as_ard(
  df_nested,
  variables,
  statistics,
  new_col_name = "..ard_all_stats..",
  omit_na = TRUE
)
}
\arguments{
\item{df_nested}{a nested data frame}

\item{variables}{character vector of variables}

\item{statistics}{named list of statistical functions}

\item{new_col_name}{string of new column name}

\item{omit_na}{logical indicating whether to omit NA values before calculating
statistics. Default is TRUE}
}
\value{
data frame
}
\description{
Calculate statistics and return in an ARD format
}
\keyword{internal}
