% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_dichotomous.R
\name{.check_dichotomous_value}
\alias{.check_dichotomous_value}
\title{Perform Value Checks}
\usage{
.check_dichotomous_value(data, value, call = get_cli_abort_call())
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{value}{(named \code{list})\cr
a named list}

\item{call}{(\code{environment})\cr
frame for error messaging. Default is \code{\link[=get_cli_abort_call]{get_cli_abort_call()}}.}
}
\value{
returns invisible if check is successful, throws an error message if not.
}
\description{
Check the validity of the values passed in \code{ard_dichotomous(value)}.
}
\examples{
cards:::.check_dichotomous_value(mtcars, list(cyl = 4))
}
\keyword{internal}
