% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_dichotomous.R
\name{.check_dichotomous_values}
\alias{.check_dichotomous_values}
\title{Perform Value Checks}
\usage{
.check_dichotomous_values(data, values, call = parent.frame())
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{values}{(named \code{list})\cr
a named list}

\item{call}{(\code{environment})\cr
frame for error messaging. Default is \code{\link[=parent.frame]{parent.frame()}}.}
}
\value{
returns invisible if check is successful, throws an error message if not.
}
\description{
Check the validity of the values passed in \code{ard_dichotomous(values)}.
}
\examples{
cards:::.check_dichotomous_values(mtcars, list(cyl = 4))
}
\keyword{internal}
