% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_statistic_fmt_fn.R
\name{.check_fmt_string}
\alias{.check_fmt_string}
\title{Check 'xx' Format Structure}
\usage{
.check_fmt_string(x, call = caller_env())
}
\arguments{
\item{x}{(\code{string} scalar)\cr
string to check}

\item{call}{(\code{environment})\cr
frame for error messaging. Default is \code{\link[=caller_env]{caller_env()}}}
}
\value{
a logical
}
\description{
A function that checks a \strong{single} string for consistency.
String must begin with 'x' and only consist of x's, a single period or none,
and may end with a percent symbol.

If string is consistent, \code{TRUE} is returned. Otherwise an error.
}
\examples{
cards:::.check_fmt_string("xx.x") # TRUE
cards:::.check_fmt_string("xx.x\%") # TRUE
}
\keyword{internal}
