% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical.R
\name{.check_for_missing_combos_in_denom}
\alias{.check_for_missing_combos_in_denom}
\title{Check for missing levels in \code{denominator}}
\usage{
.check_for_missing_combos_in_denom(
  data,
  denominator,
  by,
  strata,
  env = caller_env()
)
}
\arguments{
\item{data}{data frame}

\item{denominator}{denominator data frame}

\item{by}{character vector of by column names}

\item{strata}{character vector of strata column names}

\item{env}{environment for error messaging}
}
\value{
invisible
}
\description{
When a user passes a data frame in the \code{denominator} argument, this function
checks that the data frame contains all the same levels of the \code{by}
and \code{strata} variables that appear in the \code{data}
}
\keyword{internal}
