% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.check_var_nms}
\alias{.check_var_nms}
\title{Check variable names}
\usage{
.check_var_nms(x, vars_protected)
}
\arguments{
\item{x}{data frame}

\item{vars_protected}{character vector of protected names}
}
\value{
data frame
}
\description{
Check variable names in a data frame against protected names and modifies
them if needed
}
\keyword{internal}
