% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.check_var_nms}
\alias{.check_var_nms}
\title{Check Variable Names}
\usage{
.check_var_nms(x, vars_protected)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame}

\item{vars_protected}{(\code{character})\cr
a character vector of protected names}
}
\value{
a data frame
}
\description{
Checks variable names in a data frame against protected names and modifies
them if needed
}
\examples{
data <- data.frame(a = "x", b = "y", c = "z", .cards_idx = 1)

cards:::.check_var_nms(data, vars_protected = c("x", "z"))
}
\keyword{internal}
