% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ard_statistics.R
\name{.create_list_for_attributes}
\alias{.create_list_for_attributes}
\title{Create List for Attributes}
\usage{
.create_list_for_attributes(ard_subset, attributes, i)
}
\arguments{
\item{ard_subset}{(\code{data.frame})\cr
an ARD data frame of class 'card'}

\item{attributes}{(\code{character})\cr
a character vector of attribute names}

\item{i}{(\code{integer})\cr
a row index number}
}
\value{
a named list
}
\description{
Create List for Attributes
}
\examples{
ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")

cards:::.create_list_for_attributes(ard, c("group1", "group1_level"), 1)
}
\keyword{internal}
