% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.default_fmt_fn}
\alias{.default_fmt_fn}
\title{Add Default Formatting Functions}
\usage{
.default_fmt_fn(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}
}
\value{
a data frame
}
\description{
Add Default Formatting Functions
}
\examples{
ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1") |>
  dplyr::mutate(fmt_fn = NA)

cards:::.default_fmt_fn(ard)
}
\keyword{internal}
