% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.fill_grps_from_variables}
\alias{.fill_grps_from_variables}
\title{Back fill group variables}
\usage{
.fill_grps_from_variables(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
This function back fills the values of group variables using
variable/variable_levels. The back filling will occur if the value of the
\code{variable} column matches the name of a grouping variable, and the grouping
variable's value is NA.
}
\keyword{internal}
