% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_named_list}
\alias{.is_named_list}
\title{Named list predicate}
\usage{
.is_named_list(x, allow_df = FALSE)
}
\arguments{
\item{x}{object to check}
}
\description{
A predicate function whether input is a named list and \emph{not} a data frame.
}
\keyword{internal}
