% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.lst_results_as_df}
\alias{.lst_results_as_df}
\title{Prepare Results as Data Frame}
\usage{
.lst_results_as_df(x, variable, fun_name)
}
\arguments{
\item{x}{named list, the result from \code{eval_capture_conditions()}}

\item{variable}{string, variable name of the results}

\item{fun_name}{string, name of function called to get results in \code{x}}
}
\value{
data frame
}
\description{
Function take the results from \code{eval_capture_conditions()}, which is a
named list, e.g. \code{list(result=, warning=, error=)}, and converts it to a data
frame.
}
\keyword{internal}
