% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.message_or_error}
\alias{.message_or_error}
\title{Message or error}
\usage{
.message_or_error(
  msg,
  error = FALSE,
  call = get_cli_abort_call(),
  envir = rlang::current_env()
)
}
\arguments{
\item{msg}{(scalar \code{character})\cr
Error message}

\item{error}{(scalar \code{logical})\cr
If this should produce an error or a warning. FALSE by default}
}
\value{
Invisible NULL
}
\description{
Either error or message depending on input
}
\examples{
\dontrun{
cards:::..warn_or_error("This will be a message", FALSE)
cards:::..warn_or_error("This will be an error", TRUE)
}
}
\keyword{internal}
