% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nested_list.R
\name{.one_row_ard_to_nested_list}
\alias{.one_row_ard_to_nested_list}
\title{Convert One Row to Nested List}
\usage{
.one_row_ard_to_nested_list(x)
}
\arguments{
\item{x}{an ARD data frame with one row}
}
\value{
expression that represents an element of a nested list
}
\description{
Convert One Row to Nested List
}
\examples{
ard_continuous(mtcars, variables = mpg) |>
  dplyr::filter(dplyr::row_number() \%in\% 1L) |>
  apply_statistic_fmt_fn() |>
  cards:::.one_row_ard_to_nested_list()
}
\keyword{internal}
