% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous.R
\name{.process_nested_list_as_df}
\alias{.process_nested_list_as_df}
\title{Convert Nested Lists to Column}
\usage{
.process_nested_list_as_df(x, arg, new_column, unlist = FALSE)
}
\arguments{
\item{x}{result data frame}

\item{arg}{the nested list}

\item{new_column}{string, new column name}

\item{unlist}{logical, whether to fully unlist final results}
}
\value{
a data frame
}
\description{
Some arguments, such as the \code{stat_label}, are passed as nested lists. This
function properly unnests these lists and adds the to the results data frame.
}
\keyword{internal}
