% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.purrr_list_flatten}
\alias{.purrr_list_flatten}
\title{A list_flatten()-like function}
\usage{
.purrr_list_flatten(x)
}
\arguments{
\item{x}{a named list}
}
\description{
Function operates similarly to \code{purrr::list_flatten(x, name_spec = "{inner}")}
}
\keyword{internal}
