% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_hierarchical.R
\name{.rename_last_group_as_variable}
\alias{.rename_last_group_as_variable}
\title{Rename Last Group to Variable}
\usage{
.rename_last_group_as_variable(df_result)
}
\arguments{
\item{df_result}{(\code{data.frame})\cr
an ARD data frame of class 'card'}
}
\value{
an ARD data frame of class 'card'
}
\description{
In the \verb{ard_hierarchical*()} functions, the last grouping variable is
renamed to variable and variable_level before being returned.
}
\keyword{internal}
