% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical.R
\name{.table_as_df}
\alias{.table_as_df}
\title{Results from \code{table()} as Data Frame}
\usage{
.table_as_df(
  data,
  variable = NULL,
  by = NULL,
  strata = NULL,
  count_column = "...ard_n..."
)
}
\arguments{
\item{data}{a data frame}

\item{variable}{a string indicating a column in data}

\item{by}{a character vector indicating columns in data}

\item{strata}{a character vector indicating columns in data}
}
\value{
data frame
}
\description{
Takes the results from \code{table()} and returns them as a data frame.
After the \code{table()} results are made into a data frame, all the variables
are made into character columns, and the function also restores the
column types to their original classes. For \code{strata} columns,
only observed combinations are returned.
}
\keyword{internal}
