% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_ard.R
\name{.trim_ard}
\alias{.trim_ard}
\title{Trim ARD}
\usage{
.trim_ard(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame}
}
\value{
data frame
}
\description{
This function ingests an ARD object and trims columns and rows for downstream use in displays. The resulting data frame contains only numeric results, no supplemental information about errors/warnings, and unnested list columns.
}
\keyword{internal}
