% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{eval_hand}
\alias{eval_hand}
\title{Evaluate a poker hand}
\usage{
eval_hand(h)
}
\arguments{
\item{h}{a \code{data.frame} of 5 cards.}
}
\value{
string hand rank
}
\description{
Evaluate the rank category of a five card poker hand.
}
\details{
Uses \code{\link[=order]{order()}} and \code{\link[=rle]{rle()}} to identify flushes, straights, pairs, three, and four of a kind.
}
\examples{
deck <- new_deck()
(hand <- deal_hand(deck))
eval_hand(hand)
}
