% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck.R
\name{eval_hand}
\alias{eval_hand}
\title{Evaluate a poker hand}
\usage{
eval_hand(hand)
}
\arguments{
\item{hand}{a hand of cards.}
}
\value{
string hand rank
}
\description{
Evaluate the rank category of a five card poker hand.
}
\examples{
hand <- deal_hand(new_deck())
print_hand(hand)
eval_hand(hand)
}
