% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_table_simple.R
\name{plan_simple}
\alias{plan_simple}
\alias{table_plan_simple_continuous}
\alias{table_plan_simple_categorical}
\title{Simple ARD Table Plans}
\usage{
table_plan_simple_continuous(
  ard,
  statistics = c("{N}", "{mean} ({sd})", "{min}, {max}")
)

table_plan_simple_categorical(ard, statistics = "{n} ({p}\%)")
}
\arguments{
\item{ard}{a single ARD object for one variable}

\item{statistics}{a character vector of statistics to include in
resulting table. Default is \code{c("{N}", "{mean} ({sd})",  "{min}, {max}")}}
}
\value{
data frame
}
\description{
Plans to convert simple ARDs to formatted data frames.
These data frames will eventually be stacked into a final table.
}
\examples{
ard_continuous(mtcars, by = cyl, variables = "mpg") |>
  table_plan_simple_continuous()

ard_categorical(mtcars, by = cyl, variables = "am") |>
  table_plan_simple_categorical()
}
\keyword{internal}
