% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.card}
\alias{print.card}
\title{Print}
\usage{
\method{print}{card}(x, n = NULL, ...)
}
\arguments{
\item{x}{object of class 'card'}

\item{n}{integer specifying the number of rows to print}

\item{...}{not used}
}
\value{
data frame
}
\description{
Print method for objects of class 'card'
}
\examples{
ard_categorical(ADSL, variables = AGEGR1) |>
  print()
}
