% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.card}
\alias{print.card}
\title{Print}
\usage{
\method{print}{card}(x, n = NULL, columns = c("auto", "all"), n_col = 6L, ...)
}
\arguments{
\item{x}{(\code{data.frame})\cr
object of class 'card'}

\item{n}{(\code{integer})\cr
integer specifying the number of rows to print}

\item{columns}{(\code{string})\cr
string indicating whether to print a selected number of columns or all.}

\item{n_col}{(\code{integer})\cr
some columns are removed when there are more than a threshold of
columns present. This argument sets that threshold. This is only used
when \code{columns='auto'} and default is \code{6L}.
Columns \code{'error'}, \code{'warning'}, \code{'context'}, and \code{'fmt_fn'} \emph{may} be removed
from the print. All other columns will be printed, even if more than \code{n_col}
columns are present.}

\item{...}{(\code{\link[rlang:dyn-dots]{dynamic-dots}})\cr
not used}
}
\value{
an ARD data frame of class 'card' (invisibly)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Print method for objects of class 'card'
}
\examples{
ard_categorical(ADSL, variables = AGEGR1) |>
  print()
}
