% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ard_conditions.R
\name{print_ard_conditions}
\alias{print_ard_conditions}
\title{Print ARD Condition Messages}
\usage{
print_ard_conditions(x, call = NULL)
}
\arguments{
\item{x}{(\code{card})\cr
an ARD data frame of class 'card'}

\item{call}{(\code{environment})\cr
if supplied, the calling function will be used in messaging to users.
Default is \code{NULL}.}
}
\description{
Function parses the errors and warnings observed while calculating the
statistics requested in the ARD and prints them to the console as messages.
}
\examples{
ard_continuous(
  ADSL,
  by = ARM,
  variables = AGE,
  statistics = ~ list(
    mean = \(x) mean(x),
    mean_warning = \(x) {
      warning("warn1")
      warning("warn2")
      mean(x)
    },
    err_fn = \(x) stop("'tis an error")
  )
) |>
  print_ard_conditions()
}
