% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_ard_columns.R
\name{rename_ard_columns}
\alias{rename_ard_columns}
\title{Rename ARD Columns}
\usage{
rename_ard_columns(x, col, col_lev = NULL)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame}

\item{col}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Name of column containing the variable names}

\item{col_lev}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Name of column containing the variable levels. If \code{NULL} (default), this will be assumed to match the \code{col} argument with \verb{_level} appended to the end (e.g. if \code{col} = "group1", \code{col_lev} will be assumed to be "group1_level")}
}
\value{
data frame
}
\description{
This function combines a pair of \code{group}/\code{group_level} or \code{variable}/\code{variable_level} columns into a
single column. The \code{group_level} or \code{variable_level} column is renamed according to the value of
the \code{group} or \code{variable} column, respectively.
}
\examples{
data <- data.frame(group1 = "A", group1_level = "B", variable = "C", variable_level = "D")

cards::rename_ard_columns(data, "group1")
cards::rename_ard_columns(data, "variable")
}
