% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{selectors}
\alias{selectors}
\alias{all_ard_groups}
\alias{all_ard_variables}
\title{ARD Selectors}
\usage{
all_ard_groups(variables = TRUE, levels = TRUE)

all_ard_variables(variables = TRUE, levels = TRUE)
}
\arguments{
\item{variables}{logical indicating whether to select the columns with the
variable names. Default is \code{TRUE}}

\item{levels}{logical indicating whether to select the columns with the
variable levels. Default is \code{TRUE}}
}
\value{
tidyselect output
}
\description{
These selection helpers match variables according to a given pattern.
\itemize{
\item \code{all_ard_groups()}: Use this function in dplyr selecting environments, such
as \code{dplyr::select()}. Function selects grouping columns, e.g. columns
named \code{"group##"} or \code{"group##_level"}.
\item \code{all_ard_variables()}: Use this function in dplyr selecting environments, such
as \code{dplyr::select()}. Function selects variables columns, e.g. columns
named \code{"variable"} or \code{"variable_level"}.
}
}
\examples{
ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")

get_ard_statistics(ard, group1_level \%in\% "Placebo", variable_level \%in\% "65-80")
}
