% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_ard_hierarchical.R
\name{sort_ard_hierarchical}
\alias{sort_ard_hierarchical}
\title{Sort Stacked Hierarchical ARDs}
\usage{
sort_ard_hierarchical(x, sort = "descending")
}
\arguments{
\item{x}{(\code{card})\cr
a stacked hierarchical ARD of class \code{'card'} created using \code{\link[=ard_stack_hierarchical]{ard_stack_hierarchical()}}.}

\item{sort}{(\code{string})\cr
type of sorting to perform. Value must be one of:
\itemize{
\item \code{"alphanumeric"} - within each hierarchical section of the ARD, rows are ordered alphanumerically (i.e. A to Z)
by \code{variable_label} text.
\item \code{"descending"} - within each hierarchical section of the ARD, count sums are calculated for each row and rows are
sorted in descending order by sum. If \code{sort = "descending"}, the \code{n} statistic is used to calculate row sums if
included in \code{statistic} for all variables, otherwise \code{p} is used. If neither \code{n} nor \code{p} are present in \code{x} for
all variables, an error will occur.
}

Defaults to \code{"descending"}.}
}
\value{
an ARD data frame of class 'card'
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr

This function is used to sort stacked hierarchical ARDs.
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ard_stack_hierarchical(
  ADAE,
  variables = c(AESOC, AEDECOD),
  by = TRTA,
  denominator = ADSL |> dplyr::rename(TRTA = ARM),
  id = USUBJID
) |>
  sort_ard_hierarchical()

ard_stack_hierarchical_count(
  ADAE,
  variables = c(AESOC, AEDECOD),
  by = TRTA,
  denominator = ADSL |> dplyr::rename(TRTA = ARM)
) |>
  sort_ard_hierarchical("descending")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=filter_ard_hierarchical]{filter_ard_hierarchical()}}
}
