% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_functions}
\alias{summary_functions}
\alias{categorical_variable_summary_fns}
\alias{continuous_variable_summary_fns}
\alias{missing_variable_summary_fns}
\title{Summary Functions}
\usage{
categorical_variable_summary_fns(summaries = c("n", "p", "N"))

continuous_variable_summary_fns(
  summaries = c("N", "mean", "sd", "median", "p25", "p75", "min", "max")
)

missing_variable_summary_fns(
  summaries = c("N_obs", "N_miss", "N_nonmiss", "p_miss", "p_nonmiss")
)
}
\arguments{
\item{summaries}{a character vector of results to include in output.
\itemize{
\item \code{categorical_variable_summary_fns()}: Select one or more from 'N', 'n', 'p'
\item \code{continuous_variable_summary_fns()}: Select one or more from
'N', 'mean', 'sd', 'median', 'p25', 'p75', 'min', 'max'.
\item \code{missing_variable_summary_fns()}: Select one or more from
'N_obs', 'N_miss', 'N_nonmiss', 'p_miss', 'p_nonmiss'.
}}
}
\value{
named list of summary functions
}
\description{
\itemize{
\item \code{continuous_variable_summary_fns()} returns a named list of summary functions
for continuous variables. Some functions include slight modifications to
their base equivalents. For example, the \code{min()} and \code{max()} functions
return \code{NA} instead of \code{Inf} when an empty vector is passed.
\item \code{categorical_variable_summary_fns()} summary functions for categorical
variables.
\item \code{missing_variable_summary_fns()} summary functions suitable for variable-level
summaries, such as number and rate of missing data.
}
}
\examples{
# continuous variable summaries
ard_continuous(
  ADSL,
  variables = "AGE",
  statistics = ~continuous_variable_summary_fns(c("N", "median"))
) |>
  flatten_ard()

# categorical variable summaries
ard_categorical(
  ADSL,
  variables = "AGEGR1",
  statistics = ~categorical_variable_summary_fns(c("n", "N"))
) |>
  flatten_ard()

# summary for rates of missing data
ard_missing(
  ADSL,
  variables = c("AGE", "AGEGR1"),
  statistics = ~missing_variable_summary_fns()
) |>
  flatten_ard()
}
