% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_as_ard.R
\name{tidy_as_ard}
\alias{tidy_as_ard}
\title{Build ARD from tidier}
\usage{
tidy_as_ard(
  lst_tidy,
  tidy_result_names,
  fun_args_to_record,
  formals = list(),
  passed_args = list(),
  context,
  lst_ard_columns
)
}
\arguments{
\item{lst_tidy}{list of tidied results constructed with \code{eval_capture_conditions()},
e.g. \code{eval_capture_conditions(t.test(mtcars$mpg ~ mtcars$am) |> broom::tidy())}}

\item{tidy_result_names}{character vector of column names expected from the
tidier method. This is used to construct blank results in the event of an error.}

\item{fun_args_to_record}{character vector of function argument names that
are added to the ARD.}

\item{formals}{the results from \code{formals()}, e.g. \code{formals(fisher.test)}.
This is used to get the default argument values from unspecified arguments.}

\item{passed_args}{named list of additional arguments passed to the modeling
function.}

\item{context}{a string added as the context column in the resulting ARD}

\item{lst_ard_columns}{named list of values that will be added to the ARD
data frame.}
}
\value{
a data frame
}
\description{
Function converts a model's one-row tidy data frame into an ARD structure.
The tidied data frame must have been constructed with
\code{eval_capture_conditions()}.
}
