% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ard.R
\name{update_ard}
\alias{update_ard}
\alias{update_ard_fmt_fn}
\alias{update_ard_stat_label}
\title{Update ARDs}
\usage{
update_ard_fmt_fn(x, variables = everything(), stat_names, fmt_fn)

update_ard_stat_label(x, variables = everything(), stat_names, stat_label)
}
\arguments{
\item{x}{(\code{data.frame})\cr
an ARD data frame of class 'card'}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables in \code{x$variable} to apply update. Default is \code{everything()}.}

\item{stat_names}{(\code{character})\cr
character vector of the statistic names (i.e. values from \code{x$stat_name}) to
apply the update.}

\item{fmt_fn}{(\code{function})\cr
a function or alias recognized by \code{alias_as_fmt_fn()}.}

\item{stat_label}{(\code{function})\cr
a string of the updated statistic label.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Functions used to update ARD formatting functions and statistic labels.
}
\examples{
ard_continuous(ADSL, variables = AGE) |>
  update_ard_fmt_fn(stat_names = c("mean", "sd"), fmt_fn = 8L) |>
  update_ard_stat_label(stat_names = c("mean", "sd"), stat_label = "Mean (SD)") |>
  apply_fmt_fn()
}
